% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{ah2rh}
\alias{ah2rh}
\title{Converts absolute humidity to relative humidity}
\usage{
ah2rh(ah, temp)
}
\arguments{
\item{temp}{temperature [K]}

\item{rh}{absolute humidity [kg/m^3]}
}
\value{
relative humidity [percent]
}
\description{
Calculates absolute humidity from relative humidity and temperature
}
