% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing-real-time.R
\name{apply_quality_control}
\alias{apply_quality_control}
\title{Apply quality control on high-frequency data (e.g. as post-processing for MRD or quadrant analysis)}
\usage{
apply_quality_control(
  u,
  v,
  w,
  temp,
  h2o = NULL,
  co2 = NULL,
  ch4 = NULL,
  do_despiking = TRUE,
  despike_u = c(-15, 15, 10, 2, 8),
  despike_v = c(-15, 15, 10, 2, 8),
  despike_w = c(-4, 4, 10, 2, 8),
  despike_temp = c(230, 300, 10, 2, 8),
  despike_h2o = c(0, 12, 10, 2, 8),
  despike_co2 = c(300, 500, 10, 4, 10),
  despike_ch4 = c(0, 12, 10, 2, 8),
  do_double_rotation = TRUE,
  do_planar_fit = FALSE,
  do_detrending = FALSE,
  do_flagging = TRUE
)
}
\arguments{
\item{u}{u-wind [m/s] (sonic)}

\item{v}{v-wind [m/s] (sonic)}

\item{w}{w-wind [m/s] (sonic)}

\item{temp}{temperature [K] (sonic)}

\item{h2o}{H2O mixing ratio (gas analyzer, optional)}

\item{co2}{CO2 mixing ratio (gas analyzer, optional)}

\item{ch4}{CH4 mixing ratio (gas analyzer, optional)}

\item{do_despiking}{logical, should the data be despiked? default \code{TRUE}}

\item{despike_u}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_u=c(-15,15,10,2,8)}}

\item{despike_v}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_v=c(-15,15,10,2,8)}}

\item{despike_w}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_w=c(-4,4,10,2,8)}}

\item{despike_temp}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_temp=c(230,300,10,2,8)}}

\item{despike_h2o}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_h2o=c(0,12,10,2,8)}}

\item{despike_co2}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_co2=c(0,12,10,2,8)}}

\item{despike_ch4}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_ch4=c(0,12,10,2,8)}}

\item{do_double_rotation}{logical, should the wind data be double rotated? default \code{TRUE}}

\item{do_planar_fit}{logical, should the data be rotated with planar fit? default \code{FALSE} (either double rotation or planar fit can be \code{TRUE})}

\item{do_detrending}{logical, should the data be linearly detrended? default \code{FALSE}}

\item{do_flagging}{logical, should the data be flagged? default \code{TRUE}, i.e. several flags are calculated, but no data is removed, can be used for quality analysis}
}
\value{
quality checked data in the same dimensions as the input variables
}
\description{
Applies quality control and rotation to high-frequency data (and outputs the high-frequency data)
}
