% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_character}
\alias{assert_character}
\title{Assets that the input is character.}
\usage{
assert_character(character, null.ok = FALSE)
}
\arguments{
\item{character}{Input that is checked for being a string}

\item{null.ok}{If true, will not assert null value}
}
\description{
Assets that the input is character.
}
