% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_file_exists}
\alias{assert_file_exists}
\title{Asserts that the input is an existing file.}
\usage{
assert_file_exists(filename)
}
\arguments{
\item{filename}{Input that is checked for being an existing file.}
}
\description{
Asserts that the input is an existing file.
}
