% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_file_is_json}
\alias{assert_file_is_json}
\title{Asserts that the input is an existing json file}
\usage{
assert_file_is_json(filename)
}
\arguments{
\item{filename}{Input that is chcked for being an existing json file.}
}
\description{
Asserts that the input is an existing json file
}
