% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_integer}
\alias{assert_integer}
\title{Asserts that input is integer}
\usage{
assert_integer(number, lower = -Inf, upper = Inf)
}
\arguments{
\item{number}{Input that is checked for being an integer.}

\item{lower}{Lower bound}

\item{upper}{Upper bound}
}
\description{
Asserts that input is integer
}
