% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_is_integer}
\alias{assert_is_integer}
\title{Asserts that input is integer}
\usage{
assert_is_integer(number, lower = -Inf, upper = Inf)
}
\arguments{
\item{number}{Input that is checked for being an integer.}
}
\description{
Asserts that input is integer
}
