% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_is_string}
\alias{assert_is_string}
\title{Asserts that the input is string.}
\usage{
assert_is_string(text)
}
\arguments{
\item{text}{Input that is checked for being a string.}
}
\description{
Asserts that the input is string.
}
