% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{averaging}
\alias{averaging}
\title{accumulating / averaging}
\usage{
averaging(var, tres1 = 0.05, tres2 = c(1, 10, 30) * 60)
}
\arguments{
\item{var}{timeseries}

\item{tres1}{time resolution [s] of the given timeseries \code{var}, default \code{tres1 = 0.05} (for 20 Hz)}

\item{tres2}{desired time resolution(s) [s] of the averaged timeseries (scalar or vector), default \code{tres2 = c(1,10,30)*60} (for 1, 10 and 30 minutes)}
}
\value{
list containing mean and standard deviation of the timeseries for the desired time interval(s)
}
\description{
averaging of a timeseries
}
\examples{
ts=rnorm(30*60*20) #30 minutes of 20 Hz measurements
averaging(ts)

}
