% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{binning}
\alias{binning}
\title{discrete binning}
\usage{
binning(var1, var2, bins)
}
\arguments{
\item{var1}{vector, variable that should be binned}

\item{var2}{vector, variable used for the binning}

\item{bins}{vector, providing the intervals of the bins of var2}
}
\value{
matrix of dimension (length(bins)-1,4) with columns representing mean, median, 25%-quantile, 74%-quantile
}
\description{
discrete binning of a variable var1 based on another variable var2 (e.g., the stability parameter zeta)
}
