% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrant-analysis.R
\name{calc_2point_cor}
\alias{calc_2point_cor}
\title{Two-point Correlation Function}
\usage{
calc_2point_cor(ts1, ts2)
}
\arguments{
\item{ts1}{timeseries at point 1 (high-resolution)}

\item{ts2}{timeseries at point 2 (high-resolution)}
}
\value{
scalar giving the two-point correlation R(ts1,ts2)
}
\description{
Two-point Correlation Function
}
\examples{
set.seed(5)
ts1=rnorm(100)
ts2=rnorm(100)
cor_2point=calc_2point_cor(ts1,ts2)

}
