% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anisotropy.R
\name{calc_anisotropy}
\alias{calc_anisotropy}
\title{Invariant analysis of Reynolds stress tensor}
\usage{
calc_anisotropy(a11, a12, a13, a22, a23, a33)
}
\arguments{
\item{a11}{R11 element of Reynolds stress tensor (scalar or vector)}

\item{a12}{R12 element of Reynolds stress tensor (scalar or vector)}

\item{a13}{R13 element of Reynolds stress tensor (scalar or vector)}

\item{a22}{R22 element of Reynolds stress tensor (scalar or vector)}

\item{a23}{R23 element of Reynolds stress tensor (scalar or vector)}

\item{a33}{R33 element of Reynolds stress tensor (scalar or vector)}
}
\value{
list containing eta, xi, xb, yb (eta,xi are the coordinates of the Lumley triangle and xb,yb the coordinates of the barycentric map)
}
\description{
Invariant analysis of Reynolds stress tensor, calculation of anisotropy
}
\examples{
calc_anisotropy(1,0,0,1,0,1) #isotropic
calc_anisotropy(1,0,1,1,0,1) #some anisotropy

}
