% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_blh}
\alias{calc_blh}
\title{Boundary Layer Height}
\usage{
calc_blh(L, ustar, f)
}
\arguments{
\item{L}{Obukhov length [m]}

\item{ustar}{friction velocity [m/s]}

\item{f}{Coriolis parameter [1/s] (e.g. from \code{calc_coriolis})}
}
\value{
Boundary layer height estimation [m]
}
\description{
Calculates boundary layer height estimate following Nieuwstadt, 1981
}
\examples{
calc_blh(-1,0.5,10^(-4))

}
