% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{calc_CH}
\alias{calc_CH}
\title{Diffusivity of heat}
\usage{
calc_CH(Ri, z = 10, z0 = 0.001, z0H = NULL, method = "LTG")
}
\arguments{
\item{Ri}{bulk Richardson number [-]}

\item{z}{measurement height [m]}

\item{z0}{roughness length for momentum [m]}

\item{z0H}{roughness length for heat [m], default \code{z0H=NULL} for using z0H = z0/10 (as in HARMONIE-AROME)}

\item{method}{defining from which paper the stability correction function should be used, details see function \code{calc_FH}}
}
\value{
CH, diffusivity coefficient of heat
}
\description{
calculates heat diffusivity
}
\examples{
calc_CH(0.1)

}
