% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{calc_circular_mean}
\alias{calc_circular_mean}
\title{Calculates circular mean}
\usage{
calc_circular_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{input vector, e.g. wind directions [degree]}

\item{na.rm}{should NA values be removed? default \code{TRUE}}
}
\value{
circular mean of x values
}
\description{
calculates circular mean
}
\examples{
wd=c(280,90)
calc_circular_mean(wd)

}
