% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{calc_CM}
\alias{calc_CM}
\title{Diffusivity of momentum}
\usage{
calc_CM(Ri, z = 10, z0 = 0.001, method = "LTG")
}
\arguments{
\item{Ri}{bulk Richardson number [-]}

\item{z}{measurement height [m]}

\item{z0}{roughness length for momentum [m]}

\item{method}{defining from which paper the stability correction function should be used, details see function \code{calc_FM}}
}
\value{
CM, momentum diffusivity coefficient
}
\description{
calculates momentum diffusivity
}
\examples{
calc_CM(0.1)

}
