% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_coriolis}
\alias{calc_coriolis}
\title{Coriolis parameter}
\usage{
calc_coriolis(phi)
}
\arguments{
\item{phi}{latitude [deg]}
}
\value{
Coriolis parameter [1/s]
}
\description{
Calculates Coriolis parameter from latitude
}
\examples{
calc_coriolis(45)

}
