% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{calc_cov}
\alias{calc_cov}
\title{Calculates covariance of two timeseries using pair-wise complete observations}
\usage{
calc_cov(x, y)
}
\arguments{
\item{x}{timeseries 1}

\item{y}{timeseries 2}
}
\value{
cov(x,y)
}
\description{
Calculates cov(x,y)
}
\examples{
set.seed(5)
x=rnorm(100)
y=rnorm(100)
y[1:10]=NA
cov_xy=calc_cov(x,y)

}
