% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{calc_csi}
\alias{calc_csi}
\title{Clear Sky Index (CSI)}
\usage{
calc_csi(temp, lw_in, rh = NULL, e = NULL)
}
\arguments{
\item{temp}{scalar or vector, temperature [K]}

\item{lw_in}{scalar or vector, longwave incoming radiation [W/m^2]}

\item{rh}{scalar or vector, relative humidity [percent]}

\item{e}{scalar or vector, vapor pressure [Pa]}
}
\value{
CSI, clear sky index
}
\description{
Calculates clear sky index
}
