% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_decoupling_metric}
\alias{calc_decoupling_metric}
\title{Decoupling metric (Omega)}
\usage{
calc_decoupling_metric(w_sd, N, z = 2)
}
\arguments{
\item{w_sd}{standard deviation of vertical velocity [m/s]}

\item{N}{Brunt-Vaisala frequency [1/s]}

\item{z}{measurement height [m]}
}
\value{
decoupling metric (Omega) [-]
}
\description{
Calculates the decoupling metric (Omega) from Peltola et al., 2021 (without vegetation)
}
\examples{
calc_decoupling_metric(1,1)

}
