% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_topography.R
\name{calc_DEV}
\alias{calc_DEV}
\title{Calculates the 'deviation from mean elevation' (DEV)}
\usage{
calc_DEV(xi, yi, elevi, x, y, elev, dmax = 1000, mode = "cartesian")
}
\arguments{
\item{xi}{x-coodinate of point of interest}

\item{yi}{y-coodinate of point of interest}

\item{elevi}{elevation of point of interest}

\item{x}{x-coordinates (2-dimensional field)}

\item{y}{y-coordinates (2-dimensional field)}

\item{elev}{elevations (2-dimensional field)}

\item{dmax}{maximum fetch distance considered}

\item{mode}{either \code{mode="cartesian"} (default) or \code{mode="lonlat"} depending on whether (x, y) are in cartesian coordinates or representing longitude, latitude}
}
\value{
DEV (with DEV>0 indicating exposure and DEV<0 shelter, i.e. elevation lower than everage)
}
\description{
Calculates the 'deviation from mean elevation' (DEV) (direction-independent) as TPI/SD (normalized)
}
