% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulence-quantities.R
\name{calc_dshear}
\alias{calc_dshear}
\title{Directional Shear}
\usage{
calc_dshear(cov_uw, cov_vw)
}
\arguments{
\item{cov_uw}{covariance cov(u,w)}

\item{cov_vw}{covariance cov(v,w)}
}
\value{
angle that describes the impact of directional shear [deg]
}
\description{
Calculates a measure for directional shear alpha_uw = arctan(cov(v,w)/cov(u,w))
}
\examples{
calc_dshear(-0.5,0) #no shear
calc_dshear(-0.5,-0.1)

}
