% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_ef}
\alias{calc_ef}
\title{Evaporative fraction}
\usage{
calc_ef(sh, lh)
}
\arguments{
\item{sh}{sensible heat flux [W/m^2]}

\item{lh}{latent heat flux [W/m^2]}
}
\value{
evaporative fraction [-]
}
\description{
Calculates the evaporative fraction EF := LH/(SH+LH)
}
\examples{
calc_ef(50,20)

}
