% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_ekman_layer_depth}
\alias{calc_ekman_layer_depth}
\title{Ekman layer thickness}
\usage{
calc_ekman_layer_depth(Km, f)
}
\arguments{
\item{Km}{eddy diffusivity [m^2/s]}

\item{f}{Coriolis parameter [1/s] (e.g. from \code{calc_coriolis})}
}
\value{
Ekman layer thickness [m] (derived from boundary layer equations)
}
\description{
Calculates Ekman layer thickness from eddy diffusivity and Coriolis parameter sqrt(2*Km/abs(f))
}
\examples{
calc_ekman_layer_depth(0.1,10^(-4))

}
