% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_evapotranspiration}
\alias{calc_evapotranspiration}
\title{Evapotranspiration}
\usage{
calc_evapotranspiration(lh, temp = NULL)
}
\arguments{
\item{lh}{latent heat flux [W/m^2]}

\item{temp}{temperature [K] (optional), if provided, the latent heat of vaporization is calculated temperature-dependent}
}
\value{
evapotranspiration [kg/(s*m^2)]
}
\description{
Calculates evapotranspiration from latent heat flux
}
