% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{calc_FH}
\alias{calc_FH}
\title{Stability correction function for heat (FH), used for calculating heat diffusivity}
\usage{
calc_FH(Ri, method = "LTG")
}
\arguments{
\item{Ri}{bulk Richardson number [-]}

\item{method}{defining from which paper the stability correction function should be used, default \code{method="LTG"} for the ones used in HARMONIE-AROME, other options: \code{method="L1979"} for Louis, 1979, \code{method="LTGrevised"} for Viterbo et al, 1999, \code{method="UKMO"} for UK MetOffice}
}
\value{
FH
}
\description{
stability correction function FM -- so far for stable conditions only (Ri > 0)
}
\examples{
calc_FH(0.1)
calc_FH(0.1,method="L1979")
calc_FH(-0.1,method="UKMO")

}
