% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint.R
\name{calc_flux_footprint}
\alias{calc_flux_footprint}
\title{Flux-Footprint Parametrization (FFP) according to Kljun et al., 2015}
\usage{
calc_flux_footprint(
  zm,
  u_mean = NA,
  h,
  L,
  v_sd,
  ustar,
  z0 = NA,
  contours = seq(0.9, 0.1, -0.1),
  nres = 1000
)
}
\arguments{
\item{zm}{measurement height [m]}

\item{u_mean}{mean horizontal wind speed [m/s] (alternatively you can also use z0)}

\item{h}{boundary-layer height [m]}

\item{L}{Obukhov length [m]}

\item{v_sd}{standard deviation of crosswind [m/s]}

\item{ustar}{friction velocity [m/s]}

\item{z0}{roughness length [m] (either u_mean or z0 have to be given)}

\item{contours}{which contour lines should be calculated? default: contours=seq(0.9,0.1,-0.1)}

\item{nres}{resolution (default is nres=1000)}
}
\value{

}
\description{
Calculates Flux-Footprint Parametrization (FFP) according to Kljun et al., 2015
}
