% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux-footprint.R
\name{calc_flux_footprint_climatology}
\alias{calc_flux_footprint_climatology}
\title{Flux-Footprint Climatology / Composite Flux Footprint}
\usage{
calc_flux_footprint_climatology(
  zm,
  ws_mean = NA,
  wd_mean = NA,
  blh,
  L,
  v_sd,
  ustar,
  z0 = NA,
  contours = seq(0.9, 0.1, -0.1),
  nres = 1000,
  plot = TRUE
)
}
\arguments{
\item{zm}{measurement height [m] (time vector)}

\item{ws_mean}{mean horizontal wind speed [m/s] (time vector, alternatively you can also use \code{z0})}

\item{wd_mean}{mean wind direction [deg] (time vector, optional, used for rotating the flux footprint)}

\item{blh}{boundary-layer height [m] (time vector)}

\item{L}{Obukhov length [m] (time vector)}

\item{v_sd}{standard deviation of crosswind [m/s] (time vector)}

\item{ustar}{friction velocity [m/s] (time vector)}

\item{z0}{roughness length [m] (time vector, either \code{ws_mean} or \code{z0} have to be given)}

\item{contours}{which contour lines should be calculated? default: \code{contours=seq(0.9,0.1,-0.1)}}

\item{nres}{resolution (default: \code{nres=1000})}

\item{plot}{logical, should the flux footprint be plotted? default \code{plot=TRUE}}
}
\value{
list containing desired (averaged) contours of flux footprint climatology
}
\description{
Calculates a Flux Footprint Climatology based on \code{calc_flux_footprint} utilizing the Flux-Footprint Parametrization (FFP) according to Kljun et al., 2015
}
\examples{

}
