% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux-footprint.R
\name{calc_flux_footprint_KM2001}
\alias{calc_flux_footprint_KM2001}
\title{Flux-Footprint Calculation according to Korman and Meixner, 2001}
\usage{
calc_flux_footprint_KM2001(
  zm,
  ws_mean = NA,
  wd_mean,
  L,
  v_sd,
  ustar,
  z0,
  contours = seq(0.9, 0.1, -0.1),
  nres = 1000,
  dx = 1,
  plot = TRUE
)
}
\arguments{
\item{zm}{measurement height [m]}

\item{ws_mean}{mean horizontal wind speed [m/s] (alternatively you can also use \code{z0})}

\item{wd_mean}{mean wind direction [deg] (used to rotate flux footprint, optional)}

\item{L}{Obukhov length [m]}

\item{v_sd}{standard deviation of crosswind [m/s]}

\item{ustar}{friction velocity [m/s]}

\item{z0}{roughness length [m] (either \code{ws_mean} or \code{z0} have to be given)}

\item{contours}{which contour lines should be calculated? default: \code{contours=seq(0.9,0.1,-0.1)}}

\item{nres}{domain size (default: \code{nres=1000} to get a domain ranging from -500 to 500)}

\item{dx}{resolution (default: \code{dx=1})}

\item{plot}{logical, should the flux footprint be plotted? default \code{plot=TRUE}}
}
\value{
list containing all relevant flux footprint information
}
\description{
Calculates the Flux-Footprint Parametrization (FFP) according to Korman and Meixner, 2001
}
\examples{
#unrotated (i.e. if wind direction not given):
#ffp=calc_flux_footprint_KM2001(zm=20,ws_mean=2,blh=200,L=-1.5,v_sd=0.6,ustar=0.4,contours=0.8)
#rotated (i.e. given wind direction):
#ffp=calc_flux_footprint_KM2001(zm=20,ws_mean=2,wd_mean=80,blh=200,L=-1.5,v_sd=0.6,ustar=0.4,contours=0.8)

}
