% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_flux_intermittency}
\alias{calc_flux_intermittency}
\title{Flux intermittency}
\usage{
calc_flux_intermittency(ts1, ts2 = NULL, nsub = 6000)
}
\arguments{
\item{ts1}{timeseries 1}

\item{ts2}{timeseries 2 (optional), if the flux should be calculated based on \code{ts1*ts2} (default \code{ts2=NULL}, i.e. \code{ts2} is not used)}

\item{nsub}{number of elements used for subsampling, default \code{nsub=6000}, which corrosponds to 5 minutes of measurements from 20 Hz sampled half-hour (containing 30*60*20 = 36000 measurements)}
}
\value{
flux intermittency [-]
}
\description{
Calculates flux intermittency FI = flux_sd/abs(flux) (flux_sd: sd of subsampled fluxes) following Mahrt, 1998 (similar to stationarity flag \code{flag_stationarity})
}
\examples{
set.seed(5)
ts1=rnorm(30)
ts2=rnorm(30)
calc_flux_intermittency(ts1,ts2,nsub=6) #as product
calc_flux_intermittency(ts1*ts2,nsub=6) #the same from one variable

}
