% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulence-quantities.R
\name{calc_frictionVelocity}
\alias{calc_frictionVelocity}
\title{Friction Velocity}
\usage{
calc_frictionVelocity(covar_uw, covar_vw)
}
\arguments{
\item{covar_uw}{covariance cov(u,w) [m^2/s^2]}

\item{covar_vw}{covariance cov(v,w) [m^2/s^2]}
}
\value{
friction velocity [m/s]
}
\description{
Calculates friction velocity from the covariances cov(u,w) and cov(v,w)
}
\examples{

}
