% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulence-quantities.R
\name{calc_gustfactor}
\alias{calc_gustfactor}
\title{Gust Factor}
\usage{
calc_gustfactor(ws_max, ws_mean)
}
\arguments{
\item{ws_max}{wind speed [m/s]}

\item{ws_mean}{wind speed maximum [m/s]}
}
\value{
gust factor [-]
}
\description{
Calculates gust factor G := ws_max/ws_mean
}
\examples{

}
