% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{calc_helmholtz_decomposition}
\alias{calc_helmholtz_decomposition}
\title{Helmholtz-Hodge decomposition}
\usage{
calc_helmholtz_decomposition(u, v, res = 1)
}
\arguments{
\item{u}{zonal wind field with dimension (x,y)}

\item{v}{meridional wind field with dimension (x,y)}

\item{res}{spatial resolution (assuming equidistant grid)}
}
\value{
list containing u_div, v_div, u_rot, v_rot
}
\description{
Calculates Helmholtz-Hodge decomposition of horizontal wind using a spectral FFT-based method: decomposition of horizontal wind in rotational and divergent part: (u,v) = (u_rot,v_rot) + (u_div,v_div)
}
\details{
The implementation is based on the Python version from https://github.com/shixun22/helmholtz.
}
\examples{
set.seed(5)
u=matrix(rnorm(100),ncol=10)
v=matrix(rnorm(100),ncol=10)
hd=calc_helmholtz_decomposition(u,v,100)

}
