% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_iw}
\alias{calc_iw}
\title{Vertical Turbulence Intensity Iw}
\usage{
calc_iw(w_sd, ws_mean)
}
\arguments{
\item{w_sd}{standard deviation of vertical wind (w-wind)}

\item{ws_mean}{horizontal wind speed}
}
\value{
vertical turbulence intensity [-]
}
\description{
Calculates vertical turbulence intensity Iw := w_sd/ws_mean
}
\examples{

}
