% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{calc_k2d}
\alias{calc_k2d}
\title{Calculates 2d (horizontal) wavenumber matrix from kx, ky}
\usage{
calc_k2d(kx, ky)
}
\arguments{
\item{kx}{wavenumber in x-direction}

\item{ky}{wavenumber in y-direction}
}
\value{
total spatial (horizontal) wavenumber k
}
\description{
Calculates 2d (horizontal) wavenumber matrix from kx, ky
}
\examples{
kx=c(1:10)/10
ky=c(1:8)/8
k=calc_k2d(kx,ky)

}
