% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-closures.R
\name{calc_Kh}
\alias{calc_Kh}
\title{Calculates eddy conductivity K_h = -cov(w,T)/(dT/dz)}
\usage{
calc_Kh(cov_wT, dT_dz)
}
\arguments{
\item{cov_wT}{covariance cov(w,T) [K m/s]}

\item{dT_dz}{vertical temperature gradient [K/m]}
}
\value{
eddy conductivity K_h [m^2/s]
}
\description{
Calculates eddy conductivity K_h
}
\examples{
calc_Kh(0.2,-1)

}
