% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_Km}
\alias{calc_Km}
\title{Calculates eddy viscosity K_m = -cov(u,w)/(du/dz)}
\usage{
calc_Km(cov_uw, du_dz)
}
\arguments{
\item{cov_uw}{covariance cov(u,w) [m^2/s^2]}

\item{du_dz}{vertical wind speed gradient [1/s]}
}
\value{
eddy viscosity K_m [m^2/s]
}
\description{
Calculates eddy viscosity K_m
}
\examples{
calc_Km(-0.2,2)

}
