% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiresolution-decomposition.R
\name{calc_mrd}
\alias{calc_mrd}
\title{Multiresolution Decomposition (MRD) according to Vickers and Mahrt, 2003}
\usage{
calc_mrd(var1, var2 = NULL, time_res = 0.05, plot = TRUE, ...)
}
\arguments{
\item{var1}{timeseries of a variable}

\item{var2}{timeseries of another variable to calculate the cospectrum of \code{var1} and \code{var2}, optional (default is \code{NULL})}

\item{time_res}{time resolution of the given timeseries in seconds (e.g., \code{time_res = 0.05} for 20 Hz)}

\item{plot}{logical, should the MRD spectrum be plotted? default \code{plot=TRUE}}

\item{...}{arguments passed to plot function}
}
\value{
MRD in form of data.frame with columns: index, m, scale, time, mean, median, q25, q75
}
\description{
Calculates multiresolution decomposition (MRD) according to Vickers and Mahrt, 2003
}
\examples{
series=c(1,3,2,5,1,2,1,3) #example used in Vickers and Mahrt, 2003
calc_mrd(series)

}
