% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-closures.R
\name{calc_N2}
\alias{calc_N2}
\title{Brunt-Vaisala frequency squared}
\usage{
calc_N2(T1, T2, dz)
}
\arguments{
\item{T1}{temperature at the lower level [K]}

\item{T2}{temperature at the upper level [K]}

\item{dz}{height difference of the two measurements [m]}
}
\value{
N2 [1/s^2]
}
\description{
calculates Brunt-Vaisala frequency squared (N^2)
}
