% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogive.R
\name{calc_ogive}
\alias{calc_ogive}
\title{Calculates Ogive (cumulative distribution function from cospectrum) based on MRD}
\usage{
calc_ogive(mrd, plot = TRUE, ...)
}
\arguments{
\item{mrd}{an object returned from \code{calc_mrd}}

\item{plot}{logical, should the ogive be plotted? default \code{plot=TRUE}}

\item{...}{arguments passed to plot function}
}
\value{
ogive as list containing index, m, scale, time, ogive based on mean, and ogive based on median with quantiles q25, q75
}
\description{
Calculates Ogive from MRD spectram
}
\examples{
set.seed(5)
series=rnorm(2^10)
mrd_test=calc_mrd(c(series))
ogive_test=calc_ogive(mrd_test,plot=FALSE)

}
