% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_ozmidov_scale}
\alias{calc_ozmidov_scale}
\title{Ozmidov scale (L_OZ)}
\usage{
calc_ozmidov_scale(epsilon, N)
}
\arguments{
\item{epsilon}{dissipation rate of TKE or w [m*s]}

\item{N}{Brunt-Vaisala frequency [1/s]}
}
\value{
Ozmidov length scale [m]
}
\description{
Calculates the Ozmidov length scale L_OZ = sqrt(epsilon/N^3), with epsilon: TKE dissipation rate, and N: Brunt-Vaisala frequency
}
\examples{
calc_ozmidov_scale(-5/3,1*10^-4)

}
