% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{calc_phih}
\alias{calc_phih}
\title{Calculates Phi_h}
\usage{
calc_phih(T1, T2, cov_wT, ustar, zm, dz)
}
\arguments{
\item{T1}{temperature at the lower level [K]}

\item{T2}{temperature at the upper level [K]}

\item{cov_wT}{covariance cov(w,T) [K m/s]}

\item{ustar}{friction velocity [m/s]}

\item{zm}{measurement/scaling height [m]}

\item{dz}{height difference of the two measurements [m]}
}
\value{
Phi_h
}
\description{
calculate scaling function Phi_h (for heat)
}
