% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{calc_phim}
\alias{calc_phim}
\title{Calculates Phi_m}
\usage{
calc_phim(U1, U2, ustar, zm, dz)
}
\arguments{
\item{U1}{wind speed at the lower level [m/s]}

\item{U2}{wind speed at the upper level [m/s]}

\item{ustar}{friction velocity [m/s]}

\item{zm}{measurement/scaling height [m]}

\item{dz}{height difference of the two measurements [m]}
}
\value{
Phi_m
}
\description{
calculates scaling function Phi_m (for momentum)
}
