% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling-functions.R
\name{calc_phit}
\alias{calc_phit}
\title{Calculates Phi_t}
\usage{
calc_phit(T1, T2, cov_wT, ustar, zm, dz)
}
\arguments{
\item{T1}{temperature at the lower level [K]}

\item{T2}{temperature at the upper level [K]}

\item{cov_wT}{covariance cov(w,T) [K m/s]}

\item{ustar}{friction velocity [m/s]}

\item{zm}{measurement/scaling height [m]}

\item{dz}{height difference of the two measurements [m]}
}
\value{
Phi_t
}
\description{
calculate scaling function Phi_t (for heat)
}
