% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{calc_phix}
\alias{calc_phix}
\title{Calculates Phi_x (general flux-variance relation)}
\usage{
calc_phix(x_sd, x, ustar)
}
\arguments{
\item{x_sd}{standard deviation of x}

\item{x}{variable that should be scaled, e.g. vertical flux of x with x = T or x = q}

\item{ustar}{friction velocity [m/s]}
}
\value{
Phi_x = sigma_x/xstar
}
\description{
calculates Phi_x = sigma_x/xstar (for general flux-variance relation)
}
