% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{calc_pottemp}
\alias{calc_pottemp}
\title{Potential temperature}
\usage{
calc_pottemp(temp, pres)
}
\arguments{
\item{temp}{temperature [K]}

\item{pres}{pressure [Pa]}
}
\value{
potential temperature [K]
}
\description{
Calculates potential temperature for given temperature and pressure
}
