% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_Pr}
\alias{calc_Pr}
\title{Calculates turbulent Prandtl number Pr = K_m/K_h}
\usage{
calc_Pr(K_m, K_h)
}
\arguments{
\item{K_m}{eddy viscosity [m^2/s]}

\item{K_h}{eddy conductivity [m^2/s]}
}
\value{
Prandtl number [-]
}
\description{
Calculates turbulent Prandtl number Pr
}
\examples{
calc_Pr(0.4,0.6)

}
