% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrant-analysis.R
\name{calc_quadrant_analysis}
\alias{calc_quadrant_analysis}
\title{Calculating Coherent Structures following Quadrant Analysis}
\usage{
calc_quadrant_analysis(
  xval,
  yval,
  do_normalization = TRUE,
  hole_sizes = seq(0, 10)
)
}
\arguments{
\item{xval}{values of x variable (vector)}

\item{yval}{values of y variable (vector)}

\item{do_normalization}{should the values be normalized? i.e. \code{(x-mean(x))/sd(x)}, default: \code{do_normalization=TRUE}}

\item{hole_sizes}{vector containing desired hole sizes (integers >= 0)}
}
\value{
list containing occurrence fraction and strength (calculated based on product and covariance) for all four quadrants (mathematical orientation)
}
\description{
Calculates occurrence fraction and strength of the four quadrants in the framework of quadrant analysis
}
\examples{
a=rnorm(100)
b=rnorm(100)
qa_ab=calc_quadrant_analysis(a,b)

}
