% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-closures.R
\name{calc_ri}
\alias{calc_ri}
\title{Calculates bulk Richardson number Ri}
\usage{
calc_ri(T1, T2, U1, U2, dz)
}
\arguments{
\item{T1}{temperature at the lower level [K]}

\item{T2}{temperature at the upper level [K]}

\item{U1}{wind speed at the lower level [m/s]}

\item{U2}{wind speed at the upper level [m/s]}

\item{dz}{height difference of the two measurements [m]}
}
\value{
Ri [-]
}
\description{
calculates Richardson number Ri
}
