% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-closures.R
\name{calc_rif}
\alias{calc_rif}
\title{Calculates flux Richardson number Ri_f}
\usage{
calc_rif(cov_wT, cov_uw, U1, U2, dz, T_mean = NULL)
}
\arguments{
\item{cov_wT}{covariance cov(w,T) [K m/s]}

\item{cov_uw}{covariance cov(u,w) [m^2/s^2]}

\item{U1}{wind speed at the lower level [m/s]}

\item{U2}{wind speed at the upper level [m/s]}

\item{dz}{height difference of the two measurements [m]}

\item{T_mean}{mean temperature [K] (optional, used instead of T0=273.15)}
}
\value{
Ri_f [-]
}
\description{
calculates flux Richardson number Ri_f = g/T_mean*cov(w,T)/(cov(u,w)*du/dz)
}
