% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{calc_spectrum}
\alias{calc_spectrum}
\title{Spectrum of timeseries by wrapping rbase::spectrum()}
\usage{
calc_spectrum(ts, nbins = 100, plot = TRUE)
}
\arguments{
\item{ts}{timeseries}

\item{nbins}{number of bins used to average the spectrum, default \code{nbins=100}}

\item{plot}{should the spectrum be plotted? default \code{plot=TRUE}}
}
\value{
binned spectrum
}
\description{
Calculates and plots the averaged turbulence spectrum (as wrapper of rbase::spectrum)
}
