% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{calc_spectrum1D}
\alias{calc_spectrum1D}
\title{Spectrum of timeseries}
\usage{
calc_spectrum1D(ts, tres = 0.05, nbins = 100, method = "fft", plot = TRUE, ...)
}
\arguments{
\item{ts}{timeseries}

\item{nbins}{number of bins used to average the spectrum, default \code{nbins=100}}

\item{plot}{should the spectrum be plotted? default \code{plot=TRUE}}
}
\value{
binned spectrum
}
\description{
Calculates and plots the averaged turbulence spectrum (using FFT or DCT)
}
