% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{calc_spectrum2D}
\alias{calc_spectrum2D}
\title{Spatial spectrum (2D)}
\usage{
calc_spectrum2D(
  field,
  xres = 1000,
  yres = NULL,
  nbins = NULL,
  method = "fft",
  plot = TRUE,
  ...
)
}
\arguments{
\item{field}{two-dimensional input field}

\item{xres}{spatial resolution in x-direction}

\item{yres}{spatial resolution in y-direction, default \code{yres=NULL} meaning that the field is equidistant and \code{yres=xres} is used}

\item{nbins}{number of bins used to average the spectrum, default \code{nbins=NULL}, i.e. no further binning is applied (means number of bins equals half of length of input timeseries)}

\item{method}{method used to calculate the spectrum, can be either FFT (fast Fourier transform) or DCT (discrete cosine transform), default FFT}

\item{plot}{should the spectrum be plotted? default \code{plot=TRUE}}

\item{...}{further arguments passed to plot function}
}
\value{
binned wavenumber spectrum from 2D FFT or DCT
}
\description{
Calculates and plots turbulence spectrum (in space) calculated using FFT or DCT (and optionally bins it)
}
