% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-functions.R
\name{calc_structure_function}
\alias{calc_structure_function}
\title{Structure functions}
\usage{
calc_structure_function(ts, order = 2)
}
\arguments{
\item{ts}{timeseries}

\item{order}{order of the structure function, typically d = 2, 3, 4}
}
\value{
structure function
}
\description{
Calculates the structure function of given timeseries and given order, S := < (ts[t+dt], ts[t])^order >
}
\examples{
ts=rnorm(100)
S2_ts=calc_structure_function(ts,2)

}
