% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_topography.R
\name{calc_Sx}
\alias{calc_Sx}
\title{Calculates the slope-based direction-dependent terrain parameter Sx that indicates the degree of shelter or exposure of a point of interest from a given digital elevation model, according to Winstral and Marks, 2002}
\usage{
calc_Sx(xi, yi, elevi, x, y, elev, wd, dmax = 1000, mode = "cartesian")
}
\arguments{
\item{xi}{x-coodinate of point of interest}

\item{yi}{y-coodinate of point of interest}

\item{elevi}{elevation of point of interest}

\item{x}{x-coordinates (2-dimensional field)}

\item{y}{y-coordinates (2-dimensional field)}

\item{elev}{elevations (2-dimensional field)}

\item{wd}{wind direction}

\item{dmax}{maximum fetch distance considered}

\item{mode}{either \code{mode="cartesian"} (default) or \code{mode="lonlat"} depending on whether (x, y) are in cartesian coordinates or representing longitude, latitude}
}
\value{
Sx (with Sx>0 indicating shelter and Sx<0 exposure)
}
\description{
Calculates the slope-based direction-dependent terrain parameter Sx from a digital elevation model
}
