% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{calc_T2m_diagnostic}
\alias{calc_T2m_diagnostic}
\title{Calculates T2m diagnostic from surface temperature and temperature at first model level using the Geleyn, 1988 interpolation (used in HARMONIE-AROME)}
\usage{
calc_T2m_diagnostic(z = 2, za = 10, Ta, Ts, Ri, z0, z0H = NULL, method = "LTG")
}
\arguments{
\item{z}{height [m] to be interpolated at, for T2m \code{z=2} (2 m)}

\item{za}{height of first atmospheric model level [m] (or measurement height when applied to observations)}

\item{Ta}{temperature [K] at za}

\item{Ts}{surface temperature [K]}

\item{Ri}{Richardson number [-]}

\item{z0}{roughness length for momentum [m]}

\item{z0H}{roughness length for heat [m], default \code{z0H=NULL} for using z0H = z0/10 (as in HARMONIE-AROME)}

\item{method}{method used to calculate stability correction functions FM and FH, default \code{method="LTG"}}
}
\value{
interpolated temperature at z (e.g. T2m for z = 2 m)
}
\description{
Calculates T2m diagnostic from surface temperature and temperature at first model level using the Geleyn, 1988 interpolation (used in HARMONIE-AROME)
}
\examples{
T2m=calc_T2m_diagnostic(2,12,288,285,0.1,0.001)

}
