% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{calc_tempprofile}
\alias{calc_tempprofile}
\title{Temperature profile from Monin-Obukhov similarity theory}
\usage{
calc_tempprofile(zs, Tstar, z0H = 1e-04, d = 0, zeta = 0, T0 = 273.15)
}
\arguments{
\item{zs}{scalar or vector, heights [m] at which the horizontal wind speed should be calculate}

\item{Tstar}{temperature scale [K], e.g. \code{Tstar = -cov_wT/ustar}}

\item{z0H}{surface roughness length for heat [m], default \code{z0=0.0001}}

\item{d}{displacement height [m], optional, default \code{d=0} (i.e. no displacement)}

\item{zeta}{stability parameter [-] to correct for stability effects, default \code{zeta=0} (i.e. no stability correction, resulting in classical logarithmic wind profile)}

\item{T0}{reference temperature [K], default \code{T0=273.15}}
}
\value{
data frame containing the requested heights \code{zs} and the calculated temperature [K] there
}
\description{
Calculates vertical profile of horizontal wind speed following Monin-Obukhov similarity theory
}
\examples{
zs=seq(1,100)
Tstar=0.2
T_neutral=calc_tempprofile(zs,Tstar)
T_unstable=calc_tempprofile(zs,Tstar,zeta=-0.2)
T_stable=calc_tempprofile(zs,Tstar,zeta=0.2)

}
