% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{calc_theta}
\alias{calc_theta}
\title{Potential temperature}
\usage{
calc_theta(temp, pres)
}
\arguments{
\item{temp}{temperature [K]}

\item{pres}{pressure [Pa]}
}
\value{
potential temperature [K]
}
\description{
Calculates potential temperature for given temperature and pressure
}
\examples{
calc_theta(273,70000)

}
