% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_ti}
\alias{calc_ti}
\title{Horizontal Turbulence Intensity TI}
\usage{
calc_ti(u_sd, v_sd, ws_mean)
}
\arguments{
\item{u_sd}{standard deviation of streamwise wind (u-wind)}

\item{v_sd}{standard deviation of crosswise wind (v-wind)}

\item{ws_mean}{horizontal wind speed}
}
\value{
horizontal turbulence intensity [-]
}
\description{
Calculates horizontal turbulence intensity \code{TI = sqrt(u_sd^2+v_sd^2)/ws_mean}
}
\examples{
calc_ti(1,1,3)

}
