% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_tke}
\alias{calc_tke}
\title{Turbulent Kinetic Energy}
\usage{
calc_tke(u_sd, v_sd, w_sd)
}
\arguments{
\item{u_sd}{standard deviation of u-wind [m/s]}

\item{v_sd}{standard deviation of v-wind [m/s]}

\item{w_sd}{standard deviation of w-wind [m/s]}
}
\value{
turbulent kinetic energy TKE [m^2/s^2]
}
\description{
Calculates turbulent kinetic energy (TKE) from u_sd, v_sd and w_sd
}
\examples{

}
