% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{calc_Tv}
\alias{calc_Tv}
\title{Virtual temperature}
\usage{
calc_Tv(temp, q)
}
\arguments{
\item{temp}{temperature [K]}

\item{q}{specific humidity [kg/kg]}
}
\value{
virtual temperature [K]
}
\description{
Calculates virtual temperature for given temperature and specific humidity (mixing ratio)
}
