% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_var}
\alias{calc_var}
\title{Velocity Aspect Ratio (VAR)}
\usage{
calc_var(u_sd, v_sd, w_sd)
}
\arguments{
\item{u_sd}{standard deviation of streamwise wind (u-wind)}

\item{v_sd}{standard deviation of crosswise wind (v-wind)}

\item{w_sd}{standard deviation of vertical wind (w-wind)}
}
\value{
velocity aspect ratio [-]
}
\description{
Calculates the velocity aspect ratio: VAR := sqrt(2)*w_sd/sqrt(u_sd^2+v_sd^2)
}
