% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{calc_vpd}
\alias{calc_vpd}
\title{Vapor pressure deficit (VPD)}
\usage{
calc_vpd(temp, rh)
}
\arguments{
\item{temp}{temperature [K]}

\item{rh}{relative humidity [percent]}
}
\value{
VPD, vapor pressure deficit [Pa]
}
\description{
Calculates vapor pressure deficit (VPD) from temperature and relative humidity using Arrhenius formula
}
\examples{
calc_vpd(273,70)

}
