% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_vtke}
\alias{calc_vtke}
\title{Turbulent Kinetic Energy Velocity Scale}
\usage{
calc_vtke(u_sd, v_sd, w_sd)
}
\arguments{
\item{u_sd}{standard deviation of u-wind [m/s]}

\item{v_sd}{standard deviation of v-wind [m/s]}

\item{w_sd}{standard deviation of w-wind [m/s]}
}
\value{
turbulent kinetic energy TKE [m^2/s^2]
}
\description{
Calculates the velocity scale of turbulent kinetic energy (TKE): Vtke := sqrt(TKE)
}
\examples{

}
