% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_windDirection}
\alias{calc_windDirection}
\title{Wind Direction}
\usage{
calc_windDirection(u, v)
}
\arguments{
\item{u}{u-wind [m/s]}

\item{v}{v-wind [m/s]}
}
\value{
wind direction [deg]
}
\description{
Calculates (horizontal) wind direction
}
