% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{calc_windprofile}
\alias{calc_windprofile}
\title{Wind profile from Monin-Obukhov similarity theory}
\usage{
calc_windprofile(zs, ustar, z0 = 1e-04, d = 0, zeta = 0)
}
\arguments{
\item{zs}{scalar or vector, heights [m] at which the horizontal wind speed should be calculate}

\item{ustar}{friction velocity [m/s]}

\item{z0}{surface roughness length [m], default \code{z0=0.0001} (note: it could be an option to calculate z0 from ustar with \code{ustar2z0()})}

\item{d}{displacement height [m], optional, default \code{d=0} (i.e. no displacement)}

\item{zeta}{stability parameter [-] to correct for stability effects, default \code{zeta=0} (i.e. no stability correction, resulting in classical logarithmic wind profile)}
}
\value{
data frame containing the requested heights \code{zs} and the calculated wind speed [m/s] there
}
\description{
Calculates vertical profile of horizontal wind speed following Monin-Obukhov similarity theory
}
\examples{
zs=seq(1,100)
ustar=0.2
u_neutral=calc_windprofile(zs,ustar)
u_unstable=calc_windprofile(zs,ustar,zeta=-0.2)
u_stable=calc_windprofile(zs,ustar,zeta=0.2)

}
