% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{calc_windspeed}
\alias{calc_windspeed}
\title{Wind Speed}
\usage{
calc_windspeed(u, v, w = NULL)
}
\arguments{
\item{u}{u-wind [m/s]}

\item{v}{v-wind [m/s]}

\item{w}{w-wind [m/s] (optional), default \code{w=NULL} for horizontal wind speed}
}
\value{
wind speed [m/s]
}
\description{
Calculates wind speed (2D or 3D)
}
\examples{
calc_windspeed(3,3,0.1)

}
