% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulence-quantities.R
\name{calc_windSpeed2D}
\alias{calc_windSpeed2D}
\title{Horizontal Wind Speed}
\usage{
calc_windSpeed2D(u, v)
}
\arguments{
\item{u}{u-wind [m/s]}

\item{v}{v-wind [m/s]}
}
\value{
wind speed [m/s]
}
\description{
Calculates horizontal wind speed
}
\examples{

}
