% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{calc_windSpeed3D}
\alias{calc_windSpeed3D}
\title{Wind Speed (3D)}
\usage{
calc_windSpeed3D(u, v, w)
}
\arguments{
\item{u}{u-wind [m/s]}

\item{v}{v-wind [m/s]}

\item{w}{w-wind [m/s]}
}
\value{
wind speed (3D) [m/s]
}
\description{
Calculates wind speed (3D)
}
