% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling-functions.R
\name{calc_xstar}
\alias{calc_xstar}
\title{Calculates xstar (denominator for general flux-variance relation)}
\usage{
calc_xstar(x, ustar)
}
\arguments{
\item{x}{variable that should be scaled}

\item{ustar}{friction velocity [m/s]}
}
\value{
xstar = x/ustar
}
\description{
calculates xstar = x/ustar (for general flux-variance relation)
}
