% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{count_spikes}
\alias{count_spikes}
\title{Count spikes}
\usage{
count_spikes(ts, thresholds = c(NA, NA))
}
\arguments{
\item{ts}{time series}

\item{thresholds}{vector with lower and upper threshold, e.g. c(0,10)}
}
\value{
number of spikes in timeseries (i.e. values lower than lower threshold and higher than upper threshold)
}
\description{
Counts spikes in timeseries
}
