% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{cov2cf}
\alias{cov2cf}
\title{Converts cov(co2,w) to CO2 flux}
\usage{
cov2cf(cov_co2w, rho = NULL)
}
\arguments{
\item{cov_co2w}{covariance cov(co2,w) [m/s]}

\item{rho}{density of air [kg/m^3] (optional)}
}
\value{
CO2 flux [kg/(m^2*s)]
}
\description{
Converts cov(co2,w) to CO2 flux
}
