% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{cov2sh}
\alias{cov2sh}
\title{Converts cov(w,T) to sensible heat flux SH}
\usage{
cov2sh(cov_wT, rho = NULL)
}
\arguments{
\item{cov_wT}{covariance cov(w,T) [K m/s]}

\item{rho}{density of air [kg/m^3] (optional)}
}
\value{
sensible heat flux [W/m^2]
}
\description{
Converts cov(T,w) to sensible heat flux SH
}
