% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{deaccumulate1h}
\alias{deaccumulate1h}
\title{deaccumulation}
\usage{
deaccumulate1h(dat, factor = -1/3600)
}
\arguments{
\item{dat}{vector (with dimension time) or array (with dimension x, y, time)}

\item{factor}{factor for unit and sign conversion, default: \code{factor = -1/3600} for converting hour to seconds and adapting the sign convention}
}
\value{
vector or array (same dimension as input), hourly deaccumulated
}
\description{
hourly deaccumulation, e.g. for fluxes from model output
}
