% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{density2mixingratio}
\alias{density2mixingratio}
\title{Conversion of density to mixing ratio}
\usage{
density2mixingratio(rho)
}
\arguments{
\item{rho}{density [kg/m^3]}
}
\value{
mixing ratio of the gas [kg/kg]
}
\description{
Conversion of density to mixing ratio
}
