% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{despiking}
\alias{despiking}
\title{Despiking}
\usage{
despiking(
  series,
  thresholds = c(NA, NA),
  mad_factor = 10,
  threshold_skewness = 2,
  threshold_kurtosis = 8
)
}
\arguments{
\item{series}{timeseries that shall be despiked}

\item{thresholds}{vector with two elements representing lower and upper bounds for despiking (pre-defined thresholds), 'NA' means that the respective bound is not used}

\item{mad_factor}{factor for the MAD test, default 'mad_factor = 10'}

\item{threshold_skewness}{threshold for skewness test, default 'threshold_skewness = 2'}

\item{threshold_kurtosis}{threshold for kurtosis test, default 'threshold_kurtosis = 8'}
}
\value{
despiked timeseries
}
\description{
Applies (up to) three despiking method based on 1) pre-defined thresholds, 2) median deviation (MAD) test and 3) skewness and kurtosis
}
\examples{
set.seed(5)
ts1=rnorm(100)
despiking(ts1,thresholds=c(-1,1))

ts2=rexp(1000)
despiking(ts2)

}
