% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{dt2dx_taylor}
\alias{dt2dx_taylor}
\title{Transform time (difference) to space (difference) using Taylor hypothesis}
\usage{
dt2dx_taylor(dt, ws)
}
\arguments{
\item{dt}{time (difference) [s]}

\item{ws}{wind speed [m/s]}
}
\value{
space (difference) dx [m]
}
\description{
Transform time difference to space difference using Taylor hypothesis
}
\examples{
dx=dt2dx_taylor(0.1,3)

}
