% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing-real-time.R
\name{EC_processing_realtime}
\alias{EC_processing_realtime}
\title{Eddy-covariance post-processing for near-real-time analysis}
\usage{
EC_processing_realtime(
  u,
  v,
  w,
  temp,
  h2o = NULL,
  co2 = NULL,
  ch4 = NULL,
  time_resolution = 0.05,
  time_averaging = 30,
  measurement_height = 1,
  do_despiking = TRUE,
  despike_u = c(-15, 15, 10, 2, 8),
  despike_v = c(-15, 15, 10, 2, 8),
  despike_w = c(-4, 4, 10, 2, 8),
  despike_temp = c(230, 300, 10, 2, 8),
  despike_h2o = NULL,
  despike_co2 = NULL,
  despike_ch4 = NULL,
  do_detrending = FALSE,
  do_double_rotation = TRUE,
  do_flagging = TRUE,
  dir_blocked = c(0, 0),
  do_SNDcorrection = TRUE,
  A = 7/8,
  B = 7/8,
  do_WPLcorrection = FALSE,
  store = TRUE,
  format_out = "txt",
  filename = NULL,
  meta = TRUE
)
}
\arguments{
\item{u}{u-wind [m/s] (sonic)}

\item{v}{v-wind [m/s] (sonic)}

\item{w}{w-wind [m/s] (sonic)}

\item{temp}{temperature [K] (sonic)}

\item{h2o}{H2O mixing ratio (gas analyzer, optional)}

\item{co2}{CO2 mixing ratio (gas analyzer, optional)}

\item{ch4}{CH4 mixing ratio (gas analyzer, optional)}

\item{time_resolution}{time resolution of the measurements [s], default 20 Hz = 0.05 s}

\item{time_averaging}{desired time averaging for flux calculations [min], default 30 minutes}

\item{measurement_height}{measurement height [m], only used for calculation of the stability parameter \code{zeta}}

\item{do_despiking}{logical, should the data be despiked? default \code{TRUE}}

\item{despike_u}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_u=c(-15,15,10,2,8)}}

\item{despike_v}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_v=c(-15,15,10,2,8)}}

\item{despike_w}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_w=c(-4,4,10,2,8)}}

\item{despike_temp}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_temp=c(230,300,10,2,8)}}

\item{despike_h2o}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_h2o=c(0,12,10,2,8)}}

\item{despike_co2}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_co2=c(0,12,10,2,8)}}

\item{despike_ch4}{vector containing 5 elements: lower and upper bound, MAD factor, threshold skewness, threshold kurtosis. Details see \code{?despiking}. Default \code{despike_ch4=c(0,12,10,2,8)}}

\item{do_detrending}{logical, should the data be linearly detrended? default \code{FALSE}}

\item{do_double_rotation}{logical, should the wind data be double rotated? default \code{TRUE}}

\item{do_flagging}{logical, should the data be flagged? default \code{TRUE}, i.e. several flags are calculated, but no data is removed, can be used for quality analysis}

\item{dir_blocked}{vector containing 2 elements: wind directions blocked through mast or tower, used in flow distortion flag only}

\item{do_SNDcorrection}{logical, should SND correction be applied to the buoyancy flux? default \code{TRUE}}

\item{A}{constant used in SND correction, default \code{A=7/8} for CSAT3 sonic}

\item{B}{constant used in SND correction, default \code{A=7/8} for CSAT3 sonic}

\item{do_WPLcorrection}{logical, should WPL correction be applied to gas fluxes? default \code{FALSE}}

\item{store}{logical, should the output be stored? default \code{TRUE}}

\item{format_out}{file format of the output, can be either \code{txt} or \code{rds} (for netcdf, see separate function), only used if \code{store=TRUE}}

\item{filename}{desired output filename, default \code{NULL}, the date and time of the run will be used to create a filename, only used if \code{store=TRUE}}

\item{meta}{logical, should meta data be stored? default \code{TRUE}}
}
\value{
data frame of post-processed eddy-covariance data (that is also stored in the output file by default)
}
\description{
An example for an eddy-covariance post-processing routine utilizing the functions from ec_processing.R
}
