% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing-routine.R
\name{ECprocessing_files}
\alias{ECprocessing_files}
\title{Eddy-covariance post-processing applied to list of files}
\usage{
ECprocessing_files(
  list_of_files,
  u,
  v,
  w,
  temp,
  h2o = NULL,
  co2 = NULL,
  ch4 = NULL,
  time_resolution = 0.05,
  time_averaging = 30,
  measurement_height = 1,
  do_despiking = TRUE,
  despike_u = c(-30, 30, 10, 2, 8),
  despike_v = c(-30, 30, 10, 2, 8),
  despike_w = c(-5, 5, 10, 2, 8),
  despike_temp = c(230, 300, 10, 2, 8),
  do_double_rotation = TRUE,
  do_planar_fit = FALSE,
  do_flagging = TRUE,
  dir_blocked = c(0, 0),
  do_SNDcorrection = TRUE,
  A = 7/8,
  B = 7/8,
  do_WPLcorrection = FALSE,
  format_out = "txt",
  meta = TRUE
)
}
\arguments{
\item{u}{u-wind [m/s] (sonic)}

\item{v}{v-wind [m/s] (sonic)}

\item{w}{w-wind [m/s] (sonic)}

\item{temp}{temperature [K] (sonic)}

\item{h2o}{specific humidity (gas analyzer, optional)}

\item{co2}{CO2 concentration (gas analyzer, optional)}

\item{ch4}{CH4 concentration (gas analyzer, optional)}

\item{time_resolution}{time resolution of the measurements [s], default 20 Hz = 0.05 s}

\item{time_averaging}{desired time averaging for flux calculations [min], default 30 minutes}

\item{measurement_height}{measurement height [m], only used for calculation of the stability parameter \code{zeta}}

\item{do_despiking}{locigal, should the data be despiked? default \code{TRUE}}

\item{do_double_rotation}{locigal, should the wind data be double rotated? default \code{TRUE}}

\item{do_planar_fit}{locigal, should the data be rotated with planar fit? default \code{FALSE} (either double rotation or planar fit can be \code{TRUE})}

\item{do_flagging}{locigal, should the data be flagged? default \code{TRUE}, i.e. several flags are calculated, but no data is removed, can be used for quality analysis}

\item{do_SNDcorrection}{locigal, should SND correction be applied to the buoyancy flux? default \code{TRUE}}

\item{do_WPLcorrection}{locigal, should WPL correction be applied to the density measurements of the gas analyszer? default \code{FALSE} (only applicable, if data from a gas analyzer is used)}

\item{format_out}{file format of the output, can be either "txt" or "rds" (for netcdf, see separate function)}

\item{meta}{locical, should meta data be stored? default \code{TRUE}}
}
\value{
no return
}
\description{
An eddy-covariance post-processing routine utilizing the functions from ec_processing.R
}
