% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{fftfreq}
\alias{fftfreq}
\title{FFT frequency}
\usage{
fftfreq(n, res = 1)
}
\arguments{
\item{n}{length}

\item{res}{spatial resolution (default: \code{res=1})}
}
\value{
vector of length n containing FFT frequencies
}
\description{
Returns FFT sampling frequencies (R version of the function fft.fftfreq in numpy)
}
\examples{
fftfreq(10)

}
