% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{filter_digits}
\alias{filter_digits}
\title{Filter digits from a corpus}
\usage{
filter_digits(corpus)
}
\arguments{
\item{corpus}{A dataframe of the corpus}
}
\value{
A filtered corpus
}
\description{
It does not remove numbers embedded in words such as 'hello123', '123hello', 'h3ll0'.
}
