% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{filter_non_alphanumeric}
\alias{filter_non_alphanumeric}
\title{Filter non-alphanumeric characters from a corpus}
\usage{
filter_non_alphanumeric(corpus)
}
\arguments{
\item{corpus}{A dataframe of the corpus}
}
\value{
A filtered corpus
}
\description{
Filter non-alphanumeric characters from a corpus
}
