% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.R
\name{find_closest_grid_point}
\alias{find_closest_grid_point}
\title{Find closest grid point}
\usage{
find_closest_grid_point(lons, lats, lon_loc, lat_loc)
}
\arguments{
\item{lons}{longitudes of a grid (vector or matrix) [deg]}

\item{lats}{latitudes of a grid (vector or matrix) [deg]}

\item{lon_loc}{longitude of the location [deg]}

\item{lat_loc}{latitude of the location [deg]}
}
\value{
array index of lons, lats that represents the closest grid point to lon_loc, lat_loc
}
\description{
Finds the closest grid point from a given point (lon_loc, lat_loc) to a given grid (lons,lats)
}
