% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosystem-respiration.r
\name{fit_light_response_curve}
\alias{fit_light_response_curve}
\title{Fit light-response curve}
\usage{
fit_light_response_curve(NEE, PAR)
}
\arguments{
\item{NEE}{net ecosystem respiration [mu mol CO2/(m^2 s)] (during daytime), timeseries}

\item{PAR}{PAR (or shortwave incoming radiation) [mu mol photons/(m^2 s)], timeseries corresponding to NEE}
}
\value{
summary of non-linear fit (with nls) for the activation energy using the Arrhenius type-temperature model
}
\description{
Fit light-response curve (rectangular-hyperbola) of NEE based on PAR (for daytime conditions)
}
