% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosystem-respiration.r
\name{fit_respiration_temperature_relation}
\alias{fit_respiration_temperature_relation}
\title{Fit Arrhenius-type temperature-respiration model}
\usage{
fit_respiration_temperature_relation(Reco, temp, Tref = 283.15)
}
\arguments{
\item{Reco}{ecosystem respiration [mu mol CO2/(m^2 s)] (during nighttime: Reco = NEE, CO2-flux), timeseries}

\item{temp}{temperature [K], timeseries corresponding to Reco}

\item{Tref}{reference temperature [K], default \code{T_ref=283.15}}
}
\value{
non-linear fit (with nls) for the activation energy using the Arrhenius type-temperature model
}
\description{
Fit Arrhenius-type temperature-respiration model, e.g. applied to fit for nighttime (when Reco = NEE) and use this further to estimate GPP
}
