% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogive.R
\name{fit_sigmoid_ogive}
\alias{fit_sigmoid_ogive}
\title{Fits sigmoid curve to ogive}
\usage{
fit_sigmoid_ogive(x, y)
}
\arguments{
\item{x}{x-values (time-axis) or directly an ogive from \code{calc_ogive}}

\item{y}{y-values (ogive) (if x is a vector)}
}
\value{
fit of sigmoid curve (\code{nls}-object)
}
\description{
Fits a sigmoid curve of the form L/(1+exp(-k*(x-x0)) (with: L: upper asymptote, k: slope, x0: midpoint) to an ogive
}
