% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{flag_distortion}
\alias{flag_distortion}
\title{Flow Distortion Flag and Wind Constancy Ratio}
\usage{
flag_distortion(u, v, dir_blocked, threshold_cr = 0.9)
}
\arguments{
\item{u}{u-wind (levelled sonic)}

\item{v}{v-wind (levelled sonic)}

\item{dir_blocked}{vector containing the lower and upper bound of the blocked wind sector in degrees (e.g., \code{dir_blocked = c(30,60)})}

\item{threshold_cr}{threshold for constancy ratio (default \code{threshold_cr = 0.9}, may be adapted to used data set)}
}
\value{
distortion flags (0: in full agreement with the criterion ... 2: does not fulfill the criterion)
}
\description{
Flow Distortion Flag according to Mauder et al., 2013: Wind coming from (pre-defined) directions blocked by the measurement device is flaged with 2 (for wind speeds greater than 0.1 assuming that during calm wind the wind direction is not well-defined). The wind constancy ratio is calculated to quantify the variability of horizontal wind direction according to Mahrt, 1999.
}
\examples{
flag_distortion(1,1,dir_blocked=c(30,60))
flag_distortion(1,1,dir_blocked=c(180,360))

}
