% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{flag_stationarity}
\alias{flag_stationarity}
\title{Stationarity Flag}
\usage{
flag_stationarity(var1, var2, nsub = 3000, thresholds_stationarity = c(0.3, 1))
}
\arguments{
\item{var1}{variable 1}

\item{var2}{variable 2 (same length as \code{var1}, usually either \code{var1} or \code{var2} represent vertical velocity)}

\item{nsub}{number of elements used for subsampling (\code{nsub < length(var1)})}

\item{thresholds_stationarity}{vector containing 2 elements to distinguish between flag=0 and flag=1, as well as flag=1 and flag=2, default: \code{c(0.3,1)}}
}
\value{
stationarity flags (0: in full agreement with the criterion ... 2: does not fulfill the criterion)
}
\description{
Stationarity Flag according to Foken and Wichura, 1996 based on the assumption that the covariance of two variables (\code{var1} and \code{var2}, one usually representing vertical velocity) calculated for blocks (of length \code{nsub}) does not differ to much from the total covariance
}
\examples{
set.seed(5)
ts1=rnorm(30)
ts2=rnorm(30)
flag_stationarity(ts1,ts2,nsub=6)

}
