% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{flag_w}
\alias{flag_w}
\title{Vertical Velocity Flag}
\usage{
flag_w(w, thresholds_w = c(0.1, 0.15))
}
\arguments{
\item{w}{vertical velocity}

\item{thresholds_w}{vector containing 2 elements to distinguish between flag=0 and flag=1, as well as flag=1 and flag=2, default: \code{c(0.1,0.15)}}
}
\value{
vertical velocity flags (0: in full agreement with the criterion ... 2: does not fulfill the criterion)
}
\description{
Vertical velocity flag according to Mauder et al., 2013: After rotation the vertical velocity should vanish, this flag flags high remaining vertical velocities.
}
\examples{
flag_w(0.01)

}
