% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{gapfilling}
\alias{gapfilling}
\title{gap-filling}
\usage{
gapfilling(var, nmissing = 4, method = "linear")
}
\arguments{
\item{var}{timeseries, where NA indicates missing values that should be filled}

\item{nmissing}{number of allowed missing values, default \code{nmissing = 4}}

\item{method}{interpolation method, can be either \code{method = "linear"} for linear interpolation (default) or \code{method = "constant"} for constant interpolation}
}
\value{
gap-filled timeseries
}
\description{
gap-filling of a timeseries based on linear or constant interpolation
}
\examples{
ts1=c(1,2,NA,0)
gapfilling(ts1) #1,2,1,0
gapfilling(ts1,method="constant") #1,2,2,0
gapfilling(ts1,nmissing=0) #too many missing values

}
