% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{get_amplitude_resolution}
\alias{get_amplitude_resolution}
\title{Amplitude resolution}
\usage{
get_amplitude_resolution(ts)
}
\arguments{
\item{ts}{time series}
}
\value{
number of spikes in timeseries (i.e. values lower than lower threshold and higher than upper threshold)
}
\description{
Gives amplitude resolution of time series (i.e. number of different values in time series)
}
