% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux-footprint.R
\name{get_contours_from_f2d}
\alias{get_contours_from_f2d}
\title{Get contours from 2D flux footprint matrix}
\usage{
get_contours_from_f2d(x, y, fmat, contours = seq(0, 0.9, 0.1))
}
\arguments{
\item{x}{x-coordinate (vector)}

\item{y}{y-coordinate (vector)}

\item{fmat}{2D flux footprint (matrix/array with the dimensions matching nx x ny)}

\item{contours}{which contours? default \code{contours=seq(0,0.9,0.1)}}
}
\value{
list with x- and y-coordinates of the contours
}
\description{
Calculates contours for given flux footprint
}
