% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux-footprint.R
\name{get_ffp_area}
\alias{get_ffp_area}
\title{Get flux footprint area from f2d}
\usage{
get_ffp_area(fmat, alpha = 0.8, dx = 1, dy = 1)
}
\arguments{
\item{fmat}{2D flux footprint (matrix/array from one of the \code{calc_flux_footprint()} functions)}

\item{alpha}{threshold for area calculation, default \code{alpha=0.8} for taking the 80 percent contour line as boundary for flux footprint area}

\item{dx}{x-resolution [m]}

\item{dy}{y-resolution [m]}
}
\value{
flux footprint area [m^2]
}
\description{
Calculates flux footprint area
}
