% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{lh2et}
\alias{lh2et}
\title{Evapotranspiration}
\usage{
lh2et(lh, temp = NULL)
}
\arguments{
\item{lh}{latent heat flux [W/m^2]}

\item{temp}{temperature [K] (optional), if provided, the latent heat of vaporization is calculated temperature-dependent}
}
\value{
evapotranspiration [kg/(s*m^2)]
}
\description{
Converts latent heat flux to evaporation
}
\examples{
lh2et(20)
lh2et(20,273)

}
