% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux-footprint.R
\name{locate_flux_footprint}
\alias{locate_flux_footprint}
\title{Transform flux footprint from (x,y)-coordinates to (lon,lat)-coordinates through given station location}
\usage{
locate_flux_footprint(ffp, lon_station, lat_station)
}
\arguments{
\item{ffp}{ffp object returned by \code{calc_flux_footprint}}

\item{lon_station}{lon of station location}

\item{lat_station}{lat of station location}
}
\value{
ffp object which contains ffp also in lon-lat coordinates (with extension _earth)
}
\description{
Transform flux footprint from (x,y)-coordinates to (lon,lat)-coordinates through given station location
}
\examples{
lon1=7.527061462
lat1=60.59384155
ffp=calc_flux_footprint(zm=20,ws_mean=2,blh=200,L=-1.5,v_sd=0.6,ustar=0.4,contours=0.8)
ffp=locate_flux_footprint(ffp,lon1,lat1)

}
