% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{molarconcentration2density}
\alias{molarconcentration2density}
\title{Conversion of molar concentration to density}
\usage{
molarconcentration2density(c, gas = "H2O")
}
\arguments{
\item{c}{molar concentration in mol/m^3}

\item{gas}{which gas? can be either \code{H2O}, \code{CO2}, \code{CH4}}
}
\value{
density of the gas [kg/m^3]
}
\description{
Conversion of molar concentration to density
}
