% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anisotropy.R
\name{plot_barycentric_map}
\alias{plot_barycentric_map}
\title{Plot in barycentric map}
\usage{
plot_barycentric_map(xb, yb, ...)
}
\arguments{
\item{xb}{xb coordinate (e.g., from calc_anisotropy)}

\item{yb}{yb coordinate (e.g., from calc_anisotropy)}

\item{...}{parameters passed to plot function}
}
\value{
plots (xb, yb) in barycentric map (no return)
}
\description{
Plots (xb, yb) from invariant analysis of Reynolds stress tensor (calc_anisotropy) in barycentric map
}
\examples{
example1=calc_anisotropy(rep(1,100),rep(0,100),rnorm(100,0,1),rep(1,100),rep(0,100),rep(1,100))
plot_barycentric_map(example1$xb,example1$yb)

}
