% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint.R
\name{plot_flux_footprint}
\alias{plot_flux_footprint}
\title{Plot Flux-Footprint}
\usage{
plot_flux_footprint(ffp, levels = c(0, 10^seq(-6, -3, 0.1)), ...)
}
\arguments{
\item{ffp}{an object returned from calc_flux_footprint}
}
\value{

}
\description{
Plots Flux-Footprint Parametrization (FFP) according to Kljun et al., 2015
}
\examples{
ffp=calc_flux_footprint(zm=20,z0=0.01,h=200,L=-100,v_sd=0.6,ustar=0.4,contours=seq(0.1,0.9,0.1))
plot(ffp)

}
