% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiresolution-decomposition.R
\name{plot_mrd}
\alias{plot_mrd}
\title{Plotting Multiresolution Decomposition}
\usage{
plot_mrd(
  mrd_out,
  suggest_avgtime = TRUE,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{mrd_out}{an object returned from \code{calc_mrd}}

\item{suggest_avgtime}{logical, should the suggested averaging time be plotted? default \code{TRUE}}

\item{xlab}{x-axis label (optional)}

\item{ylab}{y-axis label (optional)}

\item{ylim}{y-axis limits (optional)}

\item{...}{more arguments passed to plot function}
}
\value{
creates a plot of MRD with logarithmic time scale (no return)
}
\description{
Plots multiresolution decomposition (MRD)
}
\examples{
set.seed(5)
series=rnorm(2^10)
mrd_test=calc_mrd(c(series))
plot_mrd(mrd_test)

}
