% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{pres2height}
\alias{pres2height}
\title{Converts pressure to height (using barometric formula)}
\usage{
pres2height(pres, pres0 = 101315, temp0 = 288.15)
}
\arguments{
\item{pres}{pressure [Pa]}

\item{pres0}{reference pressure, scalar [Pa], default \code{pres0=101315}}

\item{temp0}{reference temperature, scalar [K], default \code{temp0=288.15}}
}
\value{
height [m]
}
\description{
Calculates height from pressure
}
\examples{
pres2height(60000) #using default surface values
pres2height(60000,95000,265) #adapted surface values

}
