% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC_2008_01_1000.R
\docType{data}
\name{RC_2008_01_1000}
\alias{RC_2008_01_1000}
\title{Reddit comments from 2008-01}
\format{Data frame with columns
\describe{
\item{id}{Unique identification}
\item{author}{Name of the author (deleted comments have [deleted])}
\item{subreddit}{Name of the subreddit the comment was posted in}
\item{body}{The comment body}
}}
\source{
\url{https://www.reddit.com/r/datasets/comments/3bxlg7/i_have_every_publicly_available_reddit_comment/}
}
\usage{
RC_2008_01_1000
}
\description{
1000 Reddit comments written in 2008-01.
}
\keyword{datasets}
