% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_reddit_raw}
\alias{read_reddit_raw}
\title{Reads a json file containing Reddit comments in a single pass.
Assumes that the comments are enclosed in '[]' and separated by ','.}
\usage{
read_reddit_raw(filename, ...)
}
\arguments{
\item{filename}{The filename}
}
\value{
A dataframe constaining the Reddit comments
}
\description{
Reads a json file containing Reddit comments in a single pass.
Assumes that the comments are enclosed in '[]' and separated by ','.
}
\note{
This is very memory hungry. For larger files the recommended method is \code{\link{read_reddit_stream}}.
}
