% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_reddit_stream}
\alias{read_reddit_stream}
\title{Reads a json file containing Reddit comments in the form of a stream.
It assumes that every line in the file is a comment record in the form of a json object.}
\usage{
read_reddit_stream(filename, pagesize = 10000)
}
\arguments{
\item{filename}{The filename}

\item{pagesize}{The number of rows read in each iteration}
}
\value{
A dataframe containing the Reddit comments
}
\description{
Reads a json file containing Reddit comments in the form of a stream.
It assumes that every line in the file is a comment record in the form of a json object.
}
