% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-description.R
\docType{package}
\name{Reddy package}
\alias{Reddy package}
\title{Introduction}
\description{
EC postprocessing and analysis
}
\details{
to be detailed
}
\references{
\itemize{
\item Mack, L., Berntsen, T.K., Vercauteren, N., Pirk, N. (2024). Transfer Efficiency and Organization in Turbulent Transport over Alpine Tundra. Boundary-Layer Meteorology 190, 38. doi: https://doi.org/10.1007/s10546-024-00879-5
}
}
