% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddy.R
\docType{package}
\name{reddy}
\alias{reddy}
\alias{reddy-package}
\title{reddy R package}
\description{
An R package that parses and cleans Reddit user comments aquired from sources such as \url{https://www.reddit.com/r/datasets/comments/3bxlg7/i_have_every_publicly_available_reddit_comment/}.
}
