% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{remove_attrs}
\alias{remove_attrs}
\title{Removes attributes from a corpus}
\usage{
remove_attrs(corpus, attrs = NULL, negative = FALSE)
}
\arguments{
\item{corpus}{A dataframe of the corpus}

\item{attrs}{A character with the attributes to remove}

\item{negative}{A boolean, if true then the input \code{attrs} are instead kept.}
}
\value{
A filtered corpus
}
\description{
Removes attributes from a corpus
}
