% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{remove_non_alphanumeric}
\alias{remove_non_alphanumeric}
\title{Removes non-alphanumeric characters from a corpus}
\usage{
remove_non_alphanumeric(corpus)
}
\arguments{
\item{corpus}{A dataframe of the corpus}
}
\value{
A filtered corpus
}
\description{
Removes non-alphanumeric characters from a corpus
}
