% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{rh2ah}
\alias{rh2ah}
\title{Converts relative humidity to absolute humidity}
\usage{
rh2ah(rh, temp)
}
\arguments{
\item{rh}{relative humidity [percent]}

\item{temp}{temperature [K]}
}
\value{
absolute humidity [kg/m^3]
}
\description{
Calculates absolute humidity from relative humidity and temperature
}
\examples{
rh2ah(70,273)

}
