% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{rh2q}
\alias{rh2q}
\title{Converts relative humidity to specific humidity}
\usage{
rh2q(rh, temp, pres)
}
\arguments{
\item{rh}{relative humidity [percent]}

\item{temp}{temperature [K]}

\item{pres}{pressure [Pa]}
}
\value{
specific humidity [kg/kg]
}
\description{
Calculates specific humidity from relative humidity, temperature and pressure
}
\examples{
rh2q(70,273,101300)

}
