% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{Ri2zeta}
\alias{Ri2zeta}
\title{Converts Richardson number Ri to stability parameter zeta using the relation from Holtslag et al, 2014}
\usage{
Ri2zeta(Ri, gradient_method = "bulk")
}
\arguments{
\item{Ri}{Richardson number [-] (bulk or gradient, see argument \code{gradient_method})}

\item{gradient_method}{method to calculate vertical gradient, either \code{method="bulk"} for surface to atmosphere or \code{method="gradient"} for two atmospheric levels}
}
\value{
stability parameter zeta [-]
}
\description{
converts Ri to zeta using the relation from Holtslag et al, 2014
}
\examples{
zeta_transformed=Ri2zeta(0.1)

}
