% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{RTcorrection}
\alias{RTcorrection}
\title{Response-time correction factor (spectral correction)}
\usage{
RTcorrection(cospectrum, freq, tau = 1)
}
\arguments{
\item{cospectrum}{cospectrum}

\item{freq}{frequency [Hz], corresponding to the cospectrum, i.e. same length}

\item{tau}{response time of the instrument [s] (has to be determined first by comparison with another instrument, that samples faster)}
}
\value{
response-time correction factor (which then can be used to correct the covariance and fluxes by multiplication)
}
\description{
Calculates the response-time correction factor from cospectrum, e.g. Peltola et al., 2021
}
