% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-closures.R
\name{scale_phih}
\alias{scale_phih}
\title{Scaling function for heat Phi_h}
\usage{
scale_phih(zeta, method = "BD")
}
\arguments{
\item{zeta}{stability parameter [-]}

\item{method}{defining from which paper the scaling function should be used, default \code{method="ecmwf"} for for the ones used in ECMWF-IFS}
}
\value{
Phi_h
}
\description{
scaling function Phi_h
}
