% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{scale_phim}
\alias{scale_phim}
\title{Scaling function for momentum Phi_m}
\usage{
scale_phim(zeta, method = "ecmwf")
}
\arguments{
\item{zeta}{stability parameter [-]}

\item{method}{defining from which paper the scaling function should be used, default \code{method="ecmwf"} for the ones used in ECMWF-IFS, other options: \code{method="B1971"} for Businger et al., 1971, and \code{DH1970} for Dyer and Hicks, 1970}
}
\value{
Phi_m
}
\description{
scaling function Phi_m
}
\examples{
scale_phim(-1)
scale_phim(1,method="B1971")

}
