% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{scale_phiT}
\alias{scale_phiT}
\title{Scaling function for temperature Phi_T}
\usage{
scale_phiT(zeta, method = "K1994")
}
\arguments{
\item{zeta}{stability parameter [-]}

\item{method}{defining from which paper the scaling function should be used, default \code{method="K1994"} for Katul, 1994, other option \code{method="SC2018"} for Stiperski and Calaf, 2018}
}
\value{
Phi_T
}
\description{
scaling function Phi_T
}
