% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling-functions.R
\name{scale_phiu}
\alias{scale_phiu}
\title{Scaling function for horizontal windspeed Phi_u}
\usage{
scale_phiu(zeta, method = "PD1984")
}
\arguments{
\item{zeta}{stability parameter [-]}

\item{method}{defining from which paper the scaling function is used, default \code{method="PD1984"} for Panofsky and Dutton, 1984}
}
\value{
Phi_u
}
\description{
scaling function Phi_u
}
