% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling-functions.R
\name{scale_phiw}
\alias{scale_phiw}
\title{Scaling function for vertical windspeed Phi_w}
\usage{
scale_phiw(zeta, method = "PD1984")
}
\arguments{
\item{zeta}{stability parameter [-]}

\item{method}{defining from which paper the scaling function is used, default \code{method="PD1984"} for Panofsky and Dutton, 1984}
}
\value{
Phi_w
}
\description{
scaling function Phi_w
}
