% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{shade_between}
\alias{shade_between}
\title{Shades area between two functions}
\usage{
shade_between(x1, x2, f1, f2, ...)
}
\arguments{
\item{x1}{x-coordintes of f1}

\item{x2}{x-coordintes of f2}

\item{f1}{y-coordintes of f1}

\item{f2}{y-coordintes of f2}

\item{...}{arguments passed to plot function}
}
\value{
no return
}
\description{
Shades area between two functions given in form of x1, x2, f1, f2
}
