% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{shift2maxccf}
\alias{shift2maxccf}
\title{Shifting two timeseries to match maximum cross-correlation}
\usage{
shift2maxccf(var1, var2, maxlag = 0, plot = FALSE)
}
\arguments{
\item{var1}{vector, first timeseries}

\item{var2}{vector, second timeseries}

\item{maxlag}{maximum number of lags to be shifted (number of samples, not time), optional}

\item{plot}{logical, should the cross-correlation be plotted? default \code{plot = FALSE}}
}
\value{
a matrix cotaining timeseries \code{var1} and \code{var2} as columns after shifting to the maximum cross-correlation
}
\description{
Shifts two timeseries to match their maximum cross-correlation (can be used e.g. for lag-time correction)
}
\examples{
ts1=runif(10)
ts2=c(1,1,ts1)
shifted=shift2maxccf(ts1,ts2)

}
