% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.R
\name{sigma2height}
\alias{sigma2height}
\title{Converts hybrid (terrain-following) sigma levels to physical heights}
\usage{
sigma2height(hybrid, Tv = 273.15)
}
\arguments{
\item{hybrid}{scalar or vector, hybrid sigma levels}

\item{Tv}{virtual temperature}
}
\value{
hybrid levels converted to physical height [m]
}
\description{
Converts hybrid (terrain-following) sigma levels to physical heights
}
\examples{
sigma2height(0.1)
sigma2height(0.1,288)

}
