% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{smaller_than_machine_epsilon}
\alias{smaller_than_machine_epsilon}
\title{Set everything smaller than machine epsilon to zero}
\usage{
smaller_than_machine_epsilon(vec)
}
\arguments{
\item{vec}{vector/time series}
}
\value{
vector of same length, just all values smaller than machine epsilon are set to exactly zero
}
\description{
Calculates machine epsilon (machine-dependent) and sets everything smaller to exactly zero
}
\examples{
ts=c(1,0.1,1e-15,1e-16,1e-17,1e-18,1e-19)
ts=smaller_than_machine_epsilon(ts)

}
