% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{SNDcorrection}
\alias{SNDcorrection}
\title{SND and cross-wind correction of sensible heat flux}
\usage{
SNDcorrection(
  Ts_mean,
  u_mean,
  v_mean,
  cov_uw,
  cov_vw,
  cov_wTs,
  cov_qw = NULL,
  A = 7/8,
  B = 7/8
)
}
\arguments{
\item{Ts_mean}{sonic temperature [K] (averaged)}

\item{u_mean}{u-wind [m/s] (averaged)}

\item{v_mean}{v-wind [m/s] (averaged)}

\item{cov_uw}{cov(u,w) [m^2/s^2]}

\item{cov_vw}{cov(v,w) [m^2/s^2]}

\item{cov_wTs}{cov(Ts,w) [K*m/s] (buoyancy flux)}

\item{cov_qw}{cov(q,w) [kg/kg*m/s] (optional)}

\item{A}{constant used in cross-wind correction, default \code{A = 7/8} for CSAT3}

\item{B}{constant used in cross-wind correction, default \code{B = 7/8} for CSAT3}
}
\value{
SND correction of sensible heat flux
}
\description{
SND and cross-wind correction of sensible heat flux: converts the buoyancy flux cov(w,Ts) (based on sonic temperature Ts) to sensible heat flux
}
