% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{sos2Ts}
\alias{sos2Ts}
\title{Converts speed of sound (sos) to sonic temperature}
\usage{
sos2Ts(sos)
}
\arguments{
\item{sos}{speed of sound [m/s]}
}
\value{
sonic temperature (virtual temperature) [K]
}
\description{
Converts speed of sound (sos) to sonic temperature
}
