% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiresolution-decomposition.R
\name{suggest_avgtime_from_mrd}
\alias{suggest_avgtime_from_mrd}
\title{Suggestion of Averaging Time from Multiresolution Decomposition}
\usage{
suggest_avgtime_from_mrd(mrd_out)
}
\arguments{
\item{mrd_out}{an object returned from \code{calc_mrd}}
}
\value{
vector or scalar containing all zero crossing
}
\description{
Suggested averaging time, i.e. first zero-crossing of MRD
}
\examples{
set.seed(5)
series=rnorm(2^10)
mrd_test=calc_mrd(c(series))
suggest_avgtime_from_mrd(mrd_test)

}
