% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{Ts2T}
\alias{Ts2T}
\title{Ts2T}
\usage{
Ts2T(Ts, q)
}
\arguments{
\item{Ts}{sonic temperature [K] (similar as virtual temperature)}

\item{q}{specific humidity [kg/kg]}
}
\value{
temperature [K]
}
\description{
Converts sonic temperature Ts to temperature T
}
