% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{ustar2z0}
\alias{ustar2z0}
\title{Calculates surface roughness length z0 from friction velocity using the simple estimate from Charnock, 1955}
\usage{
ustar2z0(ustar)
}
\arguments{
\item{ustar}{friction velocity [m/s]}
}
\value{
surface roughness length [m]
}
\description{
Calculates surface roughness z0 from friction velocity using the simple estimate from Charnock, 1955: z0 = alpha*ustar^2/g with alpha=0.016 and g=9.81 m/s^2
}
