% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{zeta2Ri}
\alias{zeta2Ri}
\title{Converts stability parameter zeta to Richardson number Ri using Businger-Dyer relations}
\usage{
zeta2Ri(zeta)
}
\arguments{
\item{zeta}{stability parameter [-]}
}
\value{
Richardson number [-]
}
\description{
converts zeta to Ri using Businger-Dyer relations
}
\examples{
Ri_transformed=zeta2Ri(0.1)

}
