% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{binning}
\alias{binning}
\title{discrete binning}
\usage{
binning(var1, var2, bins)
}
\arguments{
\item{var1}{vector, variable that should be binned}

\item{var2}{vector, variable used for the binning}

\item{bins}{vector, providing the intervals of the bins of \code{var2}}
}
\value{
matrix of dimension (\code{length(bins)-1,4}) with columns representing mean, median, 25%-quantile, 75%-quantile
}
\description{
discrete binning of a variable \code{var1} based on another variable \code{var2} (e.g., the stability parameter \code{zeta})
}
\examples{
zeta_bins=c(-10^(3:-3),10^(-3:3))
zeta_vals=rnorm(1000)
vals=runif(1000)
binned=binning(vals,zeta_vals,zeta_bins)

}
