% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_br}
\alias{calc_br}
\title{Bowen ratio BR}
\usage{
calc_br(sh, lh)
}
\arguments{
\item{sh}{sensible heat flux [W/m^2]}

\item{lh}{latent heat flux [W/m^2]}
}
\value{
Bowen ratio [-]
}
\description{
Calculates Bowen ratio BR := SH/LH
}
