% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{calc_csi}
\alias{calc_csi}
\title{Clear Sky Index (CSI)}
\usage{
calc_csi(temp, lw_in, rh = NULL, e = NULL)
}
\arguments{
\item{temp}{temperature [K]}

\item{lw_in}{longwave incoming radiation [W/m^2]}

\item{rh}{relative humidity [percent]}

\item{e}{vapor pressure [Pa] (either rh or e have to be given)}
}
\value{
CSI, clear sky index
}
\description{
Calculates clear sky index
}
\examples{
calc_csi(273,230,70) #with relative humidity

}
