% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.R
\name{calc_distance}
\alias{calc_distance}
\title{Calculates distance between two points given in lon,lat}
\usage{
calc_distance(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{longitude location 1 [deg]}

\item{lat1}{latitude location 1 [deg]}

\item{lon2}{longitude location 2 [deg]}

\item{lat2}{latitude location 2 [deg]}
}
\value{
distance between two points [m]
}
\description{
Calculates y-derivative for equidistant grid
}
\examples{
set.seed(5)
field=matrix(rnorm(16),ncol=4)
df_dy(field,10)

}
