% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux-footprint.R
\name{calc_flux_footprint_climatology}
\alias{calc_flux_footprint_climatology}
\title{Calculate Flux Footprint Climatology}
\usage{
calc_flux_footprint_climatology(
  zm,
  ws_mean = NA,
  wd_mean = NA,
  blh,
  L,
  v_sd,
  ustar,
  z0 = NA,
  contours = seq(0.9, 0.1, -0.1),
  nres = 1000,
  method = "Kljun2015",
  plot = TRUE
)
}
\arguments{
\item{zm}{measurement height [m]}

\item{ws_mean}{mean horizontal wind speed [m/s] (alternatively you can also use \code{z0})}

\item{wd_mean}{mean wind direction [deg] (used to rotate flux footprint, optional)}

\item{blh}{boundary-layer height [m]}

\item{L}{Obukhov length [m]}

\item{v_sd}{standard deviation of crosswind [m/s]}

\item{ustar}{friction velocity [m/s]}

\item{z0}{roughness length [m] (either \code{ws_mean} or \code{z0} have to be given)}

\item{contours}{which contour lines should be calculated? default: \code{contours=seq(0.9,0.1,-0.1)}}

\item{nres}{resolution (scalar) (default: \code{nres=1000})}

\item{method}{method used to calculate FFP (default \code{method="Kljun2015"})}

\item{plot}{logical, should the flux footprint be plotted? default \code{plot=TRUE}}
}
\value{
list containing all relevant flux footprint information
}
\description{
Calculates the Flux-Footprint Parametrization (FFP) according to Kljun et al., 2015
}
\examples{
nit=2
ffp=calc_flux_footprint_climatology(zm=20,ws_mean=rep(2,nit),blh=rep(200,nit),L=rep(-1.5,nit),
         v_sd=rep(0.6,nit),ustar=rep(0.4,nit),contours=0.8,wd_mean=c(50,240))

}
