% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{calc_L}
\alias{calc_L}
\title{Obukhov length}
\usage{
calc_L(ustar, T_mean, cov_wT)
}
\arguments{
\item{ustar}{friction velocity (e.g., from \code{calc_ustar}) [m/s]}

\item{T_mean}{mean temperature [K]}

\item{cov_wT}{covariance cov(w,T) [m/s K]}
}
\value{
Obukhov length [m]
}
\description{
Calculates Obukhov length from friction velocity, mean temperature and cov(T,w)
}
