% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quandrant-analysis.R
\name{calc_quadrant_analysis}
\alias{calc_quadrant_analysis}
\title{Calculating Coherent Structures following Quadrant Analysis}
\usage{
calc_quadrant_analysis(
  xval,
  yval,
  do_normalization = TRUE,
  hole_sizes = seq(0, 10)
)
}
\arguments{
\item{xval}{values of x variable (vector)}

\item{yval}{values of y variable (vector)}

\item{do_normalization}{should the values be normalized? i.e. (x-mean(x))/sd(x)}

\item{hole_sizes}{vector containing desired hole sizes (integers >= 0)}
}
\value{
list containing occurrence fraction and strength (calculated based on product and covariance) for all four quadrants (mathematical orientation)
}
\description{
Calculates Occurrence Fraction and Strength of the four Quadrants
}
\examples{
#TODO

}
