% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-meteorology.R
\name{calc_satvaporpressure}
\alias{calc_satvaporpressure}
\title{Saturation vapor pressure over water}
\usage{
calc_satvaporpressure(temp)
}
\arguments{
\item{temp}{temperature [K]}
}
\value{
E_s, saturation vapor pressure over water [Pa]
}
\description{
Calculates the saturation vapor pressure over water for given temperature and pressure
}
