% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{calc_spectrum1D}
\alias{calc_spectrum1D}
\title{Frequency spectrum (1D)}
\usage{
calc_spectrum1D(
  ts,
  tres = 0.05,
  nbins = NULL,
  method = "fft",
  na.rm = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{ts}{timeseries of the variable for which the spectrum should be calculated}

\item{tres}{time resolution [s] of the given timeseries, default \code{tres=0.05} for 20 Hz}

\item{nbins}{number of bins used to average the spectrum, default \code{nbins=NULL}, i.e. no further binning is applied (means number of bins equals half of length of input timeseries)}

\item{method}{method used to calculate the spectrum, can be either FFT (fast Fourier transform) or DCT (discrete cosine transform), default FFT}

\item{na.rm}{should NA values be removed from the timeseries? default \code{na.rm=TRUE}}

\item{plot}{should the spectrum be plotted? default \code{plot=TRUE}}

\item{...}{further arguments passed to plot function}
}
\value{
binned frequency spectrum from 1D FFT
}
\description{
Calculates and plots turbulence spectrum (in time) calculated using FFT (and optionally bins it)
}
\examples{
set.seed(5)
ts=rnorm(1000)
calc_spectrum1D(ts) #no binning
calc_spectrum1D(ts,nbins=100) #binning

}
