% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{calc_tempprofile}
\alias{calc_tempprofile}
\title{Temperatyre profile from Monin-Obukhov similarity theory}
\usage{
calc_tempprofile(zs, ustar, z0 = 0, d = 0, zeta = 0, T0 = 273.15)
}
\arguments{
\item{zs}{scalar or vector, heights [m] at which the horizontal wind speed should be calculate}

\item{ustar}{friction velocity [m/s]}

\item{z0}{surface roughness length [m], default \code{z0=0} (note: it could be an option to calculate z0 from ustar with \code{ustar2z0()})}

\item{d}{displacement height [m], optional, default \code{d=0} (i.e. no displacement)}

\item{zeta}{stability parameter [-] to correct for stability effects, default \code{zeta=0} (i.e. no stability correction, resulting in classical logarithmic wind profile)}

\item{T0}{reference temperature [K], default \code{T0=273.15}}
}
\value{
data frame containing the requested heights \code{zs} and the calculated temperature [K] there
}
\description{
Calculates vertical profile of horizontal wind speed following Monin-Obukhov similarity theory
}
\examples{
zs=seq(1,100)
ustar=0.2
T_neutral=calc_tempprofile(zs,ustar)
T_unstable=calc_tempprofile(zs,ustar,zeta=-0.2)
T_stable=calc_tempprofile(zs,ustar,zeta=0.2)

}
