% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{cov2cf}
\alias{cov2cf}
\title{Converts cov(w,co2) to CO2 flux}
\usage{
cov2cf(cov_co2w, rho = NULL)
}
\arguments{
\item{cov_co2w}{covariance cov(co2,w) [m/s]}

\item{rho}{density of air [kg/m^3] (optional)}
}
\value{
latent heat flux [W/m^2]
}
\description{
Converts cov(co2,w) to CO2 flux
}
