% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-turbulence.R
\name{cov2lh}
\alias{cov2lh}
\title{Converts cov(w,q) to latent heat flux LH}
\usage{
cov2lh(covar_wq, rho = NULL)
}
\arguments{
\item{rho}{density of air [kg/m^3] (optional)}

\item{cov_wT}{covariance cov(w,q) [m/s]}
}
\value{
latent heat flux [W/m^2]
}
\description{
Converts cov(q,w) to latent heat flux LH
}
