% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{flag_most}
\alias{flag_most}
\title{Integral Turbulence Characteristics Flag}
\usage{
flag_most(sigma_w, ustar, zeta)
}
\arguments{
\item{sigma_w}{standard deviation of vertical velocity}

\item{ustar}{friction velocity}

\item{zeta}{stability parameter \code{zeta = z/L}}
}
\value{

}
\description{
Integral Turbulence Characteristics Flag: Tests the consistency with Monin-Obukhov similarity theory using the scaling functions from Panofsky and Dutton, 1984.
}
\examples{
itc_flag=flag_most(0.2,0.4,-0.3)

}
