% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{flag_w}
\alias{flag_w}
\title{Vertical Velocity Flag}
\usage{
flag_w(w)
}
\arguments{
\item{w}{vertical velocity}
}
\value{
vertical velocity flags (0: in full agreement with the criterion ... 2: does not fulfill the criterion)
}
\description{
Vertical Velocity Flag according to Mauder et al., 2013: After rotation the vertical velocity should vanish, this flag flags high remaining vertical velocities.
}
\examples{
flag_w(0.01)

}
