% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{flag_w}
\alias{flag_w}
\title{Vertical Velocity Flag}
\usage{
flag_w(w)
}
\arguments{
\item{w}{vertical velocity}
}
\value{

}
\description{
Vertical Velocity Flag according to Mauder et al., 2013: After (planar fit) rotation the vertical velocity should vanish, this flag flags high remaining vertical velocities.
}
\examples{

}
