% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.R
\name{height2pres}
\alias{height2pres}
\title{Converts height ot pressure}
\usage{
height2pres(height, pres0 = 101315, temp0 = 273.15)
}
\arguments{
\item{height}{height/elevation [m]}

\item{pres0}{surface pressure [Pa], default \code{pres0=101315}}

\item{temp0}{reference temperature [k], default \code{temp0=273.15}}
}
\value{
pressure [Pa] (at the given height levels)
}
\description{
Converts physical height to pressure based on given surface pressure and temperature
}
\examples{
height2pres(c(10,100,1000),pres=99900)

}
