% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anisotropy.R
\name{plot_barycentric_map}
\alias{plot_barycentric_map}
\title{Plot in barycentric map}
\usage{
plot_barycentric_map(xb, yb, contours = c(5, 10, 20))
}
\arguments{
\item{xb}{xb coordinate (e.g., from \code{calc_anisotropy})}

\item{yb}{yb coordinate (e.g., from \code{calc_anisotropy})}

\item{contours}{vector containing levels of contour lines for 2d kernel density estimation, default: \code{contours=c(5,10,20)}}
}
\value{
plots (xb, yb) in barycentric map with 2d kernel density estimation (no return)
}
\description{
Plots \code{(xb, yb)} from invariant analysis of Reynolds stress tensor (\code{calc_anisotropy}) in barycentric map
}
\examples{
nm=100
example1=calc_anisotropy(rep(1,nm),rep(0,nm),runif(nm,0,1),
	rep(1,nm),rep(0,nm),runif(nm,1,1.5))
plot_barycentric_map(example1$xb,example1$yb)

}
