% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint.R
\name{plot_flux_footprint}
\alias{plot_flux_footprint}
\title{Plot Flux-Footprint}
\usage{
plot_flux_footprint(ffp, levels = c(0, 10^seq(-6, -3, 0.1)))
}
\arguments{
\item{ffp}{an object returned from calc_flux_footprint}

\item{levels}{levels used for filled.contour plot of footprint, default levels=c(0,10^seq(-6,-3,0.1))}
}
\value{

}
\description{
Plots Flux-Footprint Parametrization (FFP) according to Kljun et al., 2015
}
\examples{
ffp=calc_flux_footprint(zm=5,u_mean=5,h=700,L=-1.3,v_sd=1.2,ustar=0.35)
plot_flux_footprint(ffp)

}
