% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux-footprint.R
\name{plot_flux_footprint}
\alias{plot_flux_footprint}
\title{Plot Flux-Footprint}
\usage{
plot_flux_footprint(
  ffp,
  levels = c(0, 10^seq(-6, -3, 0.1)),
  mode = "distance",
  ...
)
}
\arguments{
\item{ffp}{an object returned from \code{calc_flux_footprint_[method]}}

\item{levels}{levels used for filled contour plot of footprint, default \code{levels=c(0,10^seq(-6,-3,0.1))}}

\item{mode}{can be either \code{mode="distance"} for plotting footprint relative to station location in cartesian coordinates or \code{mode="lonlat"} for plotting in (lon,lat)-ccordinates}

\item{...}{paraemters passed to image.plot function}
}
\value{
no return
}
\description{
Plots Flux-Footprint Parametrization (FFP)
}
\examples{
ffp=calc_flux_footprint_Kljun2015(zm=5,ws_mean=5,blh=700,L=-1.3,v_sd=1.2,ustar=0.35)
plot_flux_footprint(ffp)

}
