% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrd.R
\name{plot_mrd}
\alias{plot_mrd}
\title{Plotting Multiresolution Decomposition}
\usage{
plot_mrd(mrd_out, ...)
}
\arguments{
\item{mrd_out}{an output object from calc_mrd}

\item{...}{parameters passed to plot function}
}
\value{
creates a plot of MRD with logarithmic time scale (no return)
}
\description{
Plots multiresolution decomposition (MRD)
}
\examples{
set.seed(5)
series=rnorm(2^10)
mrd_test=calc_mrd(c(series))
plot_mrd(mrd_test)

}
