% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogive.R
\name{plot_ogive}
\alias{plot_ogive}
\title{Plotting Ogive}
\usage{
plot_ogive(ogive, ...)
}
\arguments{
\item{ogive}{an object returned from \code{calc_ogive}}

\item{...}{arguments passed to plot function}
}
\value{
creates a plot of an ogive with logarithmic time scale (no return)
}
\description{
Plots ogive
}
\examples{
set.seed(5)
series=rnorm(2^10)
mrd_test=calc_mrd(c(series))
ogive_test=calc_ogive(mrd_test)
plot_ogive(ogive_test)

}
