% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrant-analysis.R
\name{plot_quadrant_analysis}
\alias{plot_quadrant_analysis}
\title{Plotting Quadrant Analysis}
\usage{
plot_quadrant_analysis(
  xval,
  yval,
  do_normalization = TRUE,
  hole_sizes = c(1, 2),
  contours = 10^(-3:3),
  print_fit = TRUE,
  ...
)
}
\arguments{
\item{xval}{values of x variable (vector)}

\item{yval}{values of y variable (vector)}

\item{do_normalization}{should the values be normalized, i.e. (x-mean(x))/sd(x)? default: TRUE}

\item{hole_sizes}{vector containing desired hole sizes (integers >= 0), default: c(1,2)}

\item{contours}{vector containing levels of contour lines for 2d kernel densoty estimation, default: contours=10^(-3:3)}

\item{print_fit}{should the fit summary from the linear regression be printed? default: TRUE}

\item{...}{arguments passed to plot function}
}
\value{

}
\description{
Calculates occurrence fraction and strength of the four quadrants
}
\examples{
a=rnorm(100)
b=rnorm(100)
plot_quadrant_analysis(a,b)

}
