% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrant-analysis.R
\name{plot_quadrant_analysis}
\alias{plot_quadrant_analysis}
\title{Plotting Quadrant Analysis}
\usage{
plot_quadrant_analysis(
  xval,
  yval,
  do_normalization = TRUE,
  hole_sizes = c(1, 2, 3),
  plot_kde2d = TRUE,
  contours = 10^(-3:3),
  print_fit = TRUE,
  ...
)
}
\arguments{
\item{xval}{values of x variable (vector)}

\item{yval}{values of y variable (vector)}

\item{do_normalization}{should the values be normalized? i.e. \code{(x-mean(x))/sd(x)}, default: \code{do_normalization=TRUE}}

\item{hole_sizes}{vector containing desired hole sizes (integers >= 0), default: \code{hole_sizes=c(1,2)}}

\item{plot_kde2d}{should the contour lines of the 2d kernel density estimation be plotted? default \code{plot_kde2d = TRUE}}

\item{contours}{vector containing levels of contour lines for 2d kernel density estimation, only used if \code{plot_kde2d = TRUE}, default: \code{contours=10^(-3:3)}}

\item{print_fit}{should the fit summary from the linear regression be printed? default: \code{print_fit=TRUE}}

\item{...}{arguments passed to plot function}
}
\value{
no return
}
\description{
Plots two vectors in the framework of quadrant analysis with 2d kernel density estimation (optional)
}
\examples{
a=rnorm(100)
b=rnorm(100)
plot_quadrant_analysis(a,b)

}
