% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface-energy-balance.R
\name{plot_seb}
\alias{plot_seb}
\title{Plottting of surface energy balance and calculate surface energy balance unclosure}
\usage{
plot_seb(
  sw_in,
  sw_out,
  lw_in,
  lw_out,
  sh = NULL,
  lh = NULL,
  gh = NULL,
  time_vector = NULL,
  print_fit = TRUE,
  ...
)
}
\arguments{
\item{sw_out}{outcomin shortwave radiation [W/m^2] (as vector of time)}

\item{lw_in}{incoming longwave radiation [W/m^2] (as vector of time)}

\item{lw_out}{outgoing longwave radiation [W/m^2] (as vector of time)}

\item{sh}{sensible heat flux [W/m^2] (as vector of time) -- if measured}

\item{lh}{latent heat flux [W/m^2] (as vector of time) -- if measured}

\item{gh}{ground heat flux [W/m^2] (as vector of time) -- if measured}

\item{time_vector}{times used as x-axis labels (optional)}

\item{print_fit}{should the fit summary be printed? default: TRUE}
}
\value{
no return
}
\description{
Plottting of surface energy balance and calculate surface energy balance unclosure as residual flux and closure ratio
}
\examples{

}
