% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{ppt2rho}
\alias{ppt2rho}
\title{Unit conversion of "parts-per" to density (for closed-path gas analyzer)}
\usage{
ppt2rho(ppt, T_mean = 288.15, pres = 101325, e = 0, gas = "H2O")
}
\arguments{
\item{ppt}{measurement in parts per thousand [ppt]}

\item{T_mean}{temperature [K]}

\item{pres}{pressure [Pa]}

\item{e}{water vapor pressure [Pa]}

\item{gas}{which gas? can be either "H2O", "CO2", "CH4" (if CO2/CH4 is selected, make sure that it's still in ppt and not ppm as usual)}
}
\value{
density of the gas [kg/m^3]
}
\description{
Unit conversion of "parts-per" to density (for closed-path gas analyzer)
}
