% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{rotate_double}
\alias{rotate_double}
\title{Double rotation}
\usage{
rotate_double(u, v, w)
}
\arguments{
\item{u}{u-wind (levelled sonic)}

\item{v}{v-wind (levelled sonic)}

\item{w}{w-wind (levelled sonic)}
}
\value{
list containing the wind in a natural coordinate system (streamwise, crosswise, vertical) and the two rotation angles theta and phi
}
\description{
Double rotation
}
\examples{


}
