% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{rotate_planar}
\alias{rotate_planar}
\title{Planar fit rotation}
\usage{
rotate_planar(u, v, w, bias = c(0, 0, 0))
}
\arguments{
\item{u}{u-wind (levelled sonic)}

\item{v}{v-wind (levelled sonic)}

\item{w}{w-wind (levelled sonic)}

\item{bias}{a three-dimensional correction vector containing the offsets of u-, v-, w-wind}
}
\value{
list containing u, v, w after planar fit rotation as well as the rotation angles alpha, beta and gamma and the fitted offset c3
}
\description{
Planar fit rotation (i.e., sonic coordinate system will be aligned with the mean streamlines resulting in vanishing of w_mean)
}
\examples{
u=rnorm(1000)
v=rnorm(1000)
w=rnorm(1000)
wind_rotated=rotate_planar(u,v,w) #for planar fit a timeseries is required

}
