% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{rotate_planar}
\alias{rotate_planar}
\title{Planar fit rotation}
\usage{
rotate_planar(u, v, w, c = c(0, 0, 0))
}
\arguments{
\item{u}{u-wind (levelled sonic)}

\item{v}{v-wind (levelled sonic)}

\item{w}{w-wind (levelled sonic)}

\item{c}{a three-dimensional vector containing the offset of u-, v-, w-wind}
}
\value{

}
\description{
Planar fit rotation
}
\examples{


}
