% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{scale_phic}
\alias{scale_phic}
\title{Scaling function for scalar concentrations Phi_C}
\usage{
scale_phic(zeta, method = "")
}
\arguments{
\item{zeta}{stability parameter [-]}

\item{method}{defining from which paper the scaling function should be used ...}
}
\value{
Phi_C
}
\description{
scaling function Phi_C
}
