% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterizations-of-turbulence.R
\name{scale_phih}
\alias{scale_phih}
\title{Scaling function for heat Phi_h}
\usage{
scale_phih(zeta, method = "ecmwf")
}
\arguments{
\item{zeta}{stability parameter [-]}

\item{method}{defining from which paper the scaling function should be used, default \code{method="ecmwf"} for for the ones used in ECMWF-IFS, other options: \code{method="B1971"} for Businger et al., 1971, and \code{DH1970} for Dyer and Hicks, 1970}
}
\value{
Phi_h
}
\description{
scaling function Phi_h
}
\examples{
scale_phih(-1)
scale_phih(1,method="B1971")

}
