% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling-functions.R
\name{scale_phim}
\alias{scale_phim}
\title{Scaling function for momentum Phi_m}
\usage{
scale_phim(zeta, method = "ecmwf")
}
\arguments{
\item{zeta}{stability parameter [-]}

\item{method}{defining from which paper the scaling function should be used, default \code{method="ecmwf"} for the ones used in ECMWF-IFS, other option \code{method="BD"} for the linear Businger-Dyer relations}
}
\value{
Phi_m
}
\description{
scaling function Phi_m
}
