% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{shift2maxccf}
\alias{shift2maxccf}
\title{shifting two timeseries to match maximum cross-correlation}
\usage{
shift2maxccf(var1, var2, plot = TRUE)
}
\arguments{
\item{var1}{vector, first timeseries}

\item{var2}{vector, second timeseries}

\item{plot}{logical, should the cross-correlation be plotted? default \code{plot = TRUE}}
}
\value{
a matrix cotaining timeseries var1 and var2 as columns after shifting to the maximum cross-correlation
}
\description{
shifts two timeseries to match their maximum cross-correlation
}
\examples{
ts1=runif(10)
ts2=c(1,1,ts1)
shifted=shift2maxccf(ts1,ts2)

}
