% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{SNDcorrection}
\alias{SNDcorrection}
\title{SND correction}
\usage{
SNDcorrection(u, v, w, Ts, q = NULL, A = 7/8, B = 7/8)
}
\arguments{
\item{u}{u-wind (levelled sonic)}

\item{v}{v-wind (levelled sonic)}

\item{w}{w-wind (levelled sonic)}

\item{Ts}{temperature (sonic temperature or corrected temperature)}

\item{q}{specific humidity (if measured, default NULL)}

\item{A}{constant used in SND correction, default 'A = 7/8'}

\item{B}{constant used in SND correction, default 'B = 7/8'}
}
\value{

}
\description{
SND correction of sensible heat flux
}
\examples{

}
