% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec-processing.R
\name{WPLcorrection}
\alias{WPLcorrection}
\title{WPL correction}
\usage{
WPLcorrection(
  Ts_mean,
  q_mean,
  cov_wTs,
  rhow_mean,
  cov_wrhow,
  rhoc_mean = NULL,
  cov_wrhoc = NULL
)
}
\arguments{
\item{Ts_mean}{temperature [K] (sonic temperature or corrected temperature)}

\item{q_mean}{specific humidity [kg/kg] (if measured, default \code{NULL})}

\item{cov_wTs}{covariance cov(w,Ts) [m/s*K]}

\item{rhow_mean}{measured water vapor density [kg/m^3]}

\item{cov_wrhow}{covariance cov (w,rhow) [m/s*kg/m^3]}

\item{rhoc_mean}{measured trace gas density [kg/m^3] (only if WPL-correction should be applied to another flux, e.g. CO2 flux, default \code{NULL})}

\item{cov_wrhoc}{covariance cov (w,rhoc) [m/s*kg/m^3] (only if WPL-correction should be applied to another flux, e.g. CO2 flux, default \code{NULL})}
}
\value{
WPL correction of respective flux
}
\description{
WPL correction: density correction for trace gas fluxes (i.e., converts volume- to mass-related quantity)
}
